refactor: split workflow.py into a workflow package, one module per command
`processing/workflow.py` had grown to 3885 lines against a 1000-line
hard cap. It is now `processing/workflow/`, with a module per registered
command named after the command, which is the arrangement `hcp/` already
uses and makes the rule for finding a command "look for its name":

    workflow/
        __init__.py                  re-exports the seven commands
        dryrun.py                    the shared side-effect guards
        get_bold_data.py
        create_bold_brain_masks.py   + its execute_
        compute_bold_stats.py        + its execute_
        create_stats_report.py
        extract_nuisance_signal.py   + its execute_
        preprocess_bold.py           + its execute_
        preprocess_conc.py

**No function body changed.** The diff is a move plus one mechanical
substitution: the four dry-run helpers lose their leading underscore and
are reached through the module, so `_run_external(log, ...)` reads
`dryrun.run_external(log, ...)` at the 33 call sites that use one. The
qualifier is what now says a call honours `--test`, which is the point
of `dryrun.remove` sitting next to `os.remove`. Each command's preamble
constants move into the module of the command that uses them -- each is
used by exactly one -- and `mcommand`, read from `QUNEXMCOMMAND` at
import time and used by four of the commands, goes to `dryrun.py`
alongside them.

The helpers went to a sibling module rather than to `__init__.py`
deliberately. Defining them in the package and importing them back out
of it from the submodules does work, but only while the definitions stay
above the re-export lines: it is an import cycle held together by line
order, and moving a line would fail collection with an `ImportError`
naming the wrong module. The tree answers this twice already -- `hcp/`
keeps its shared code in `hcp_utils.py`, and `general/log/__init__.py`
is a re-export facade whose shared class lives in `report.py`.

`__init__.py` therefore re-exports the seven commands and holds nothing
else. That is what keeps `import qx_utilities.processing.workflow as wf`
working for the rest of the tree; the registry is unaffected by it,
since the builder only considers function definitions carrying a
`.. qx_command:` docstring and an import is not one.

`qx_commands.yaml` is rebuilt. The only change in it is the seven
`path:` entries, which gain their module segment -- a command's name
comes from its function, so nothing needed an alias and no option,
description or call form moved.

Verified three ways, since a refactor of this size is only worth as much
as its oracle:

- every one of the 15 functions is byte-identical to its pre-split
  source once the helper calls are substituted, checked by comparing
  AST line ranges across the two trees;
- all seven commands were driven in `--test` mode on the same fixture
  before and after, and their reports are identical once timestamps and
  the temporary path are normalised;
- `tests/test_workflow_dryrun.py` and `tests/test_mov_stats.py` pass on
  the re-export facade.

The one test edit is not a concession on that last point: the dry-run
fixture reached the `time` module through `wf.time` to stub `sleep`,
which the package does not re-export and should not. It now patches
`time.sleep` directly, which is the same object it was patching before.

Sizes, per the size targets. No module is over the 1000-line hard cap.
Four are over the 500-line preferred target: `preprocess_conc.py` (973)
and `preprocess_bold.py` (917), where 590 and 588 lines respectively are
the command's own docstring; and `create_bold_brain_masks.py` (535) and
`create_stats_report.py` (523). Reaching the target for the first two
would mean separating a command's documentation from the function it
documents, which nothing in the tree does and which would put the
registry's source of truth somewhere other than the code it describes.

Seven functions remain over the 100-line function cap, unchanged:
`preprocess_conc` (358 code lines), `execute_create_bold_brain_masks`
(338), `create_stats_report` (240), `execute_preprocess_bold` (202),
`get_bold_data` (127), `execute_extract_nuisance_signal` (125) and
`execute_compute_bold_stats` (104). Moving a function does not shorten
it, and shortening one here would have made the diff impossible to read
as a move -- which is the property the three checks above rest on. The
breach is deliberate and is left for a separate pass.

github issue: #81
