--- python/qx_utilities/processing/workflow.py
@@ -1,3905 +1,0 @@
-#!/usr/bin/env python
-# encoding: utf-8
-
-# SPDX-FileCopyrightText: 2021 QuNex development team <https://qunex.yale.edu/>
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-"""
-``workflow.py``
-
-This file holds code for running functional connectivity preprocessing and
-GLM computation workflow. It consists of functions:
-
---get_bold_data             Maps NIL preprocessed data to images folder.
---create_bold_brain_masks   Extracts the first frame of each BOLD file.
---compute_bold_stats        Computes per volume image statistics for scrubbing.
---create_stats_report       Creates a report of movement and image statistics.
---extract_nuisance_signal   Extracts the nuisance signal for regressions.
---preprocess_bold           Processes a single BOLD file.
---preprocess_conc           Processes concatenated BOLD files.
-
-All the functions are part of the processing suite. They should be called
-from the command line using `qunex` command. Help is available through:
-
-- `qunex ?<command>` for command specific help
-"""
-
-# Created by Grega Repovs on 2016-12-17.
-# Code split from dofcMRIp_core gCodeP/preprocess codebase.
-# Copyright (c) Grega Repovs. All rights reserved.
-
-import os
-import shutil
-import textwrap
-import time
-import traceback
-from concurrent.futures import ProcessPoolExecutor
-from datetime import datetime
-from functools import partial
-
-import qx_utilities.processing.core as pc
-import qx_utilities.processing.mov_stats as ms
-import qx_utilities.general.exceptions as ge
-import qx_utilities.general.filelock as fl
-import qx_utilities.general.meltmovfidl as gm
-import qx_utilities.general.img as gi
-import qx_utilities.general.core as gc
-from qx_utilities.general.log import ReportLog
-
-
-if "QUNEXMCOMMAND" not in os.environ:
-    # import time, before any command and therefore any log exists: this one
-    # stays a print (group C's frame, not group A's report)
-    print(
-        "WARNING: QUNEXMCOMMAND environment variable not set. Matlab will be run by default!"
-    )
-    mcommand = "matlab -nojvm -nodisplay -nosplash -r"
-else:
-    mcommand = os.environ["QUNEXMCOMMAND"]
-
-
-# -------------------------------------------------------- the command preambles
-#
-# What each command does, shown once at the head of every session report, and
-# the parameters it quotes back. Dedented blocks rather than lines carrying
-# their own `\n    `: this is prose, it is read as prose, and it should be
-# reviewable as prose. The parameter lists are lists rather than format strings
-# with one interpolation each, so they cannot drift from `options`.
-
-BRAIN_MASKS_PURPOSE = """\
-A mask identifying the actual coverage of the brain is created for each of the
-specified BOLD files, from its first frame. Only the images named by --bolds
-are processed; the default is all of them."""
-
-BOLD_STATS_PURPOSE = """\
-Per frame statistics are computed for each of the specified BOLD files, from
-its movement correction parameter file and an analysis of the image. The
-results are saved as *.bstat and *.bscrub files in the images/movement
-subfolder. Only the images named by --bolds are processed. Note that the NIfTI
-volume image is used even when the target format is CIFTI."""
-
-BOLD_STATS_PARAMETERS = [
-    "mov_radius",
-    "mov_fd",
-    "mov_dvars",
-    "mov_dvarsme",
-    "mov_after",
-    "mov_before",
-    "mov_bad",
-]
-
-STATS_REPORT_PURPOSE = """\
-Movement correction parameters and computed BOLD statistics are used to create
-per session plots, fidl snippets and group reports. Only the images named by
---bolds are processed; see the documentation for the other parameters."""
-
-STATS_REPORT_PARAMETERS = [
-    "mov_dvars",
-    "mov_dvarsme",
-    "mov_fd",
-    "mov_radius",
-    "mov_fidl",
-    "mov_post",
-    "mov_pref",
-]
-
-# the only one of the four that names the files it writes, so it is a template.
-# Wrapped here for review and re-wrapped after substitution -- `nifti_tail` and
-# `img_suffix` are empty as often as not, and a paragraph laid out around their
-# widest form comes out ragged when they are short
-NUISANCE_PURPOSE = """\
-Nuisance signal is extracted from each of the specified BOLD files. The results
-are saved as {boldname}[N]{nifti_tail}.nuisance files in the
-images{img_suffix}/movement subfolder. Only the images named by --bolds are
-processed. Note that the NIfTI volume image is used even when the target format
-is CIFTI."""
-
-NUISANCE_PARAMETERS = ["wbmask", "sessionroi", "nroi", "shrinknsroi"]
-
-# what the preambles are wrapped to
-PURPOSE_WIDTH = 79
-
-
-# --------------------------------------------------------------- the dry run
-#
-# Five commands in this file -- `get_bold_data`, `create_bold_brain_masks`,
-# `compute_bold_stats`, `create_stats_report` and `extract_nuisance_signal` --
-# never consulted `options["run"]`, so `--test` did the work: it copied and
-# linked files, invoked external tools, and deleted existing reports before
-# regenerating them. `preprocess_bold` and `preprocess_conc` did guard the one
-# matlab call each makes, but fell silent about it instead of naming it, and
-# `preprocess_conc` copied its conc and event files in and rewrote the conc
-# regardless of the flag.
-#
-# The four helpers below are the same guard for the side effects the commands
-# repeat, spelled once instead of at 30 sites, and they follow
-# `processing/fs.py`'s: a dry run reports what it *would* do rather than
-# falling silent, so the report is worth reading. The one-off side effects --
-# `gi.slice_image`, `os.makedirs`, `gm.meltmovfidl`, the file lock and the
-# merged comlog -- are guarded where they sit.
-
-
-def _run_external(_log, options, checkfile, command, description, **kwargs):
-    """
-    Run one external command, or -- under ``--test`` -- report it and stop.
-
-    Returns the underlying ``(endlog, status, failed)``. A dry run ran nothing
-    and wrote no comlog, so it returns ``(None, None, 0)`` -- the shape the
-    call sites unpack, rather than a bare None that would raise there.
-    """
-    if options["run"] != "run":
-        _log.raw(f"\n\n{description}")
-        _log.detail(f"test, not run: {command}", depth=1)
-        return None, None, 0
-
-    return pc.run_external_for_file(checkfile, command, description, **kwargs, _log=_log)
-
-
-def _link_or_copy(_log, options, source, target, **kwargs):
-    """Link or copy a file, or -- under ``--test`` -- report it and change nothing."""
-    if options["run"] != "run":
-        _log.detail(f"test, not copied: {os.path.basename(source)}")
-        return None
-
-    return gc.link_or_copy(source, target, **kwargs)
-
-
-def _copy(_log, options, source, target):
-    """Copy a file, or -- under ``--test`` -- report the copy and change nothing."""
-    if options["run"] != "run":
-        _log.detail(f"test, not copied: {os.path.basename(source)}")
-        return
-
-    shutil.copy2(source, target)
-
-
-def _remove(_log, options, path):
-    """
-    Delete a file, or -- under ``--test`` -- report the deletion and keep it.
-
-    Only a file that is actually there is reported. Most of these are temporary
-    files a dry run never created, and naming them would be noise; the ones
-    worth naming are the existing outputs `create_stats_report` clears before
-    regenerating them, which is the destructive step a dry run must not take.
-    """
-    if options["run"] != "run":
-        if os.path.exists(path):
-            _log.detail(f"test, not removed: {path}")
-        return
-
-    os.remove(path)
-
-
-def get_bold_data(sinfo, options, overwrite=False, thread=0):
-    """
-    ``get_bold_data [... processing options]``
-
-    Map NIL preprocessed data into the session's images folder.
-
-    ..  qx_command:
-        type: processing.session
-
-    Parameters:
-        --batchfile (str, default ''):
-            The batch.txt file with all the session information.
-
-        --sessions (str, default ''):
-            A list of sessions to process.
-
-        --sessionsfolder (str, default '.'):
-            The path to the study/sessions folder.
-
-        --overwrite (str, default 'no'):
-            Whether to overwrite existing data (yes) or not (no).
-
-        --logfolder (str, default ''):
-            The path to the folder where runlogs and comlogs are to be stored,
-            if other than default.
-    """
-    log = ReportLog()
-
-    log.rule()
-    log.info(f"Session id: {sinfo['id']} \n[started on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}]")
-    log.info("Copying imaging data ...")
-
-    log.info("Structural data ...")
-    pc.do_options_check(options, sinfo, "get_bold_data")
-    f = pc.get_file_names(sinfo, options)
-
-    with pc.combined_comlog(log, options, "get_bold_data", thread=sinfo["id"]):
-        copy = True
-        if os.path.exists(f["t1"]):
-            copy = False
-
-        try:
-            if overwrite or copy:
-                if f["t1_source"] is None:
-                    raise pc.NoSourceFolder(
-                        "ERROR: Data source folder is not set. Please check your paths!"
-                    )
-                log.detail(f"copying {f['t1_source']}")
-                if options["image_target"] == "4dfp":
-                    if gi.get_img_format(f["t1_source"]) == ".4dfp.img":
-                        _link_or_copy(log, options, f["t1_source"], f["t1"])
-                        _link_or_copy(
-                            log,
-                            options,
-                            f["t1_source"].replace(".img", ".ifh"),
-                            f["t1"].replace(".img", ".ifh"),
-                        )
-                    else:
-                        tmpfile = f["t1"].replace(
-                            ".4dfp.img", gi.get_img_format(f["t1_source"])
-                        )
-                        _link_or_copy(log, options, f["t1_source"], tmpfile)
-                        _run_external(
-                            log,
-                            options,
-                            f["t1"],
-                            "g_FlipFormat %s %s"
-                            % (tmpfile, f["t1"].replace(".img", ".ifh")),
-                            "... converting %s to 4dfp" % (os.path.basename(tmpfile)),
-                            overwrite=overwrite,
-                        )
-                        _remove(log, options, tmpfile)
-                if options["image_target"] == "nifti":
-                    if gi.get_img_format(f["t1_source"]) == ".4dfp.img":
-                        tmpimg = f["t1"] + ".4dfp.img"
-                        tmpifh = f["t1"] + ".4dfp.ifh"
-                        _link_or_copy(log, options, f["t1_source"], tmpimg)
-                        _link_or_copy(
-                            log, options, f["t1_source"].replace(".img", ".ifh"), tmpifh
-                        )
-                        _run_external(
-                            log,
-                            options,
-                            f["t1"],
-                            "g_FlipFormat %s %s"
-                            % (tmpifh, f["t1"].replace(".img", ".ifh")),
-                            "... converting %s to NIfTI" % (os.path.basename(tmpimg)),
-                            overwrite=overwrite,
-                        )
-                        _remove(log, options, tmpimg)
-                        _remove(log, options, tmpifh)
-                    else:
-                        if gi.get_img_format(f["t1_source"]) == ".nii.gz":
-                            tmpfile = f["t1"] + ".gz"
-                            _link_or_copy(log, options, f["t1_source"], tmpfile)
-                            _run_external(
-                                log,
-                                options,
-                                f["t1"],
-                                "gunzip -f %s" % (tmpfile),
-                                "... gunzipping %s" % (os.path.basename(tmpfile)),
-                                overwrite=overwrite,
-                            )
-                            if os.path.exists(tmpfile):
-                                _remove(log, options, tmpfile)
-                        else:
-                            _link_or_copy(log, options, f["t1_source"], f["t1"])
-
-            else:
-                log.detail(f"{f['t1']} present")
-        except Exception:
-            log.error("getting the data failed! Please check paths and files!", depth=1)
-
-        # the same bold selection every other command in this file uses. The
-        # hand-rolled loop this replaces matched `task` alone, split `--bolds`
-        # on "|" only, and had no case for its own default of "all" -- which
-        # matched no task and so processed nothing at all
-        bolds, _, _ = pc.use_or_skip_bold(sinfo, options, _log=log)
-
-        for boldinfo in bolds:
-            boldname = boldinfo["name"]
-
-            log.raw("\n\nWorking on: " + boldname + " ...")
-
-            try:
-                # --- filenames
-                f = pc.get_file_names(sinfo, options)
-                f.update(pc.get_bold_file_names(sinfo, boldname, options))
-                _ = pc.get_session_folders(sinfo, options)
-
-                # the bold's own data, not the status of whichever structural
-                # conversion happened to run last: that name is unbound
-                # whenever none did -- which is every run where the T1 is
-                # already in place, and every dry run -- and reading it raised
-                # a NameError the handler below reported as an unknown error,
-                # once per bold
-                if os.path.exists(f["bold_vol"]):
-                    log.step("Data ready!")
-                else:
-                    log.error("Data missing, please check source!")
-
-            except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-                log.raw(str(errormessage))
-            except Exception:
-                log.error(f"Unknown error occured: \n...................................\n{traceback.format_exc()}...................................\n")
-                time.sleep(3)
-
-    log.blank()
-    log.info(f"Imaging data copy completed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-    log.rule()
-
-    # the per-bold errors above are what the failure count is derived from
-    return log.finish("Imaging data copy completed", name=sinfo["id"])
-
-
-def create_bold_brain_masks(sinfo, options, overwrite=False, thread=0):
-    """
-    ``create_bold_brain_masks [... processing options]``
-
-    Extract the brain and create a brain mask for each BOLD image.
-
-    ..  qx_command:
-        type: processing.session
-
-    Parameters:
-        --batchfile (str, default ''):
-            The batch.txt file with all the sessions' information.
-
-        --sessionsfolder (str, default '.'):
-            The path to the study/sessions folder, where the imaging data is
-            supposed to go.
-
-        --parsessions (int, default 1):
-            How many sessions to run in parallel.
-
-        --parelements (int, default 1):
-            How many elements (e.g. bolds) to run in parallel.
-
-        --overwrite (str, default 'no'):
-            Whether to overwrite existing data (yes) or not (no). Note that
-            previous data is deleted before the run, so in the case of a failed
-            command run, previous results are lost.
-
-        --bolds (str, default 'rest'):
-            Which bold images (as they are specified in the batch.txt file) to
-            copy over. It can be a single type (e.g. 'task'), a pipe separated
-            list (e.g. 'WM|Control|rest') or 'all' to copy all.
-
-        --boldname (str, default 'bold'):
-            The default name of the bold files in the images folder.
-
-        --nifti_tail (str, default detailed below):
-            The tail of NIfTI volume images to use. Default to the value of
-            qx_nifti_tail.
-
-        --bold_variant (str, default ''):
-            Optional variant of bold preprocessing. If specified, the BOLD
-            images in `images/functional<bold_variant>` will be processed.
-
-        --img_suffix (str, default ''):
-            Specifies a suffix for 'images' folder to enable support for
-            multiple parallel workflows. Empty if not used.
-
-        --logfolder (str, default ''):
-            The path to the folder where logs are to be stored,
-            if other than default.
-
-    Notes:
-        The parameters can be specified in command call or session.txt file.
-
-        create_bold_brain_masks takes the first image of each bold file, and
-        runs FSL bet to extract the brain and create a brain mask. The resulting
-        files are saved into images/segmentation/boldmasks in the source image
-        format:
-
-        - bold[N]<nifti_tail>_frame1.*
-        - bold[N]<nifti_tail>_frame1_brain.*
-        - bold[N]<nifti_tail>_frame1_brain_mask.*
-
-    Examples:
-        ::
-
-            qunex create_bold_brain_masks \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --nifti_tail=_hp2000_clean \\
-                --bolds=all \\
-                --parelements=8
-    """
-    log = ReportLog()
-
-    report = {
-        "bolddone": 0,
-        "boldok": 0,
-        "boldfail": 0,
-        "boldmissing": 0,
-        "boldskipped": 0,
-    }
-
-    log.rule()
-    log.info(f"Session id: {sinfo['id']} \n[started on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}]")
-    log.action("Creating", "masks for bold runs ...", options["run"], level="info")
-    log.blank()
-    log.info(BRAIN_MASKS_PURPOSE)
-
-    pc.do_options_check(options, sinfo, "create_bold_brain_masks")
-    d = pc.get_session_folders(sinfo, options)
-
-    if overwrite:
-        ostatus = "will"
-    else:
-        ostatus = "will not"
-
-    log.step(f"Working on BOLD images in {d['s_images']}")
-    log.detail(f"images{options['img_suffix']}/functional{options['bold_variant']} will be processed")
-    log.detail(f"the resulting masks will be in {d['s_boldmasks']}")
-    log.detail(f"{', '.join(options['bolds'].split('|'))} BOLD files will be processed (see --bolds)")
-    log.detail(f"existing masks {ostatus} be overwritten (see --overwrite)")
-
-    bolds, bskip, report["boldskipped"] = pc.use_or_skip_bold(sinfo, options, _log=log)
-
-    parelements = options["parelements"]
-    log.info(f"Processing {parelements} BOLDs in parallel")
-
-    if parelements == 1:  # serial execution
-        for b in bolds:
-            # process
-            result = execute_create_bold_brain_masks(sinfo, options, overwrite, b)
-
-            # merge r
-            log.raw(result["r"])
-
-            # merge report
-            temp_report = result["report"]
-            report["bolddone"] += temp_report["bolddone"]
-            report["boldok"] += temp_report["boldok"]
-            report["boldfail"] += temp_report["boldfail"]
-            report["boldmissing"] += temp_report["boldmissing"]
-    else:  # parallel execution
-        # create a multiprocessing Pool
-        process_pool_executor = ProcessPoolExecutor(parelements)
-        # process
-        f = partial(execute_create_bold_brain_masks, sinfo, options, overwrite)
-        results = process_pool_executor.map(f, bolds)
-
-        # merge r and report
-        for result in results:
-            log.raw(result["r"])
-            temp_report = result["report"]
-            report["bolddone"] += temp_report["bolddone"]
-            report["boldok"] += temp_report["boldok"]
-            report["boldfail"] += temp_report["boldfail"]
-            report["boldmissing"] += temp_report["boldmissing"]
-
-    log.blank()
-    log.info(f"Bold mask creation completed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-    log.rule()
-    rstatus = (
-        "BOLDS done: %(bolddone)2d, missing data: %(boldmissing)2d, failed: %(boldfail)2d, processed: %(boldok)2d, skipped: %(boldskipped)2d"
-        % (report)
-    )
-
-    return log.result(rstatus, report["boldmissing"] + report["boldfail"], sinfo["id"])
-
-
-def execute_create_bold_brain_masks(sinfo, options, overwrite, boldinfo):
-
-    # prepare return variables
-    log = ReportLog()
-    report = {"bolddone": 0, "boldok": 0, "boldfail": 0, "boldmissing": 0}
-
-    log.step("Working on " + boldinfo["name"])
-
-    try:
-        # --- filenames
-        f = pc.get_file_names(sinfo, options)
-        f.update(pc.get_bold_file_names(sinfo, boldinfo["name"], options))
-
-        # template file
-        templatefile = f["bold_template"]
-
-        # --- copy over bold data
-        # --- bold
-        status = pc.check_for_file(f["bold_vol"],
-            "bold data present",
-            "bold data missing, skipping bold",
-            status=True,
-            _log=log,
-        )
-        if not status:
-            log.info("Looked for:" + f["bold_vol"])
-            report["boldmissing"] += 1
-            return {"r": log.text, "report": report}
-
-        # --- extract first bold frame
-        if not os.path.exists(f["bold1"]) or overwrite:
-            if options["run"] != "run":
-                # the check below reads the file the slice would have written,
-                # so a dry run has to skip it rather than report a failure and
-                # return before naming the tools it would have run
-                log.detail(f"test, not sliced: first frame of {os.path.basename(f['bold_vol'])}")
-            else:
-                gi.slice_image(f["bold_vol"], f["bold1"], 1)
-                if os.path.exists(f["bold1"]):
-                    log.detail(f"sliced first frame from {os.path.basename(f['bold_vol'])}")
-                else:
-                    log.warning(f"failed slicing first frame from {os.path.basename(f['bold_vol'])}", depth=1)
-                    report["boldfail"] += 1
-                    return {"r": log.text, "report": report}
-        else:
-            log.detail(f"first {os.path.basename(f['bold_vol'])} frame already present")
-
-        # --- logs storage
-        endlogs = []
-
-        # --- convert to NIfTI
-        bsource = f["bold1"]
-        bbtarget = f["bold1_brain"].replace(
-            gi.get_img_format(f["bold1_brain"]), ".nii.gz"
-        )
-        bmtarget = f["bold1_brain_mask"].replace(
-            gi.get_img_format(f["bold1_brain_mask"]), ".nii.gz"
-        )
-        if gi.get_img_format(f["bold1"]) == ".4dfp.img":
-            bsource = f["bold1"].replace(".4dfp.img", ".nii.gz")
-
-            # run g_FlipFormat
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                bsource,
-                "g_FlipFormat %s %s" % (f["bold1"], bsource),
-                "    ... converting %s to nifti" % (f["bold1"]),
-                overwrite=overwrite,
-                remove=options["log"] == "remove",
-                thread=sinfo["id"],
-                task="FlipFormat",
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["logtag"],
-                    "B%d" % boldinfo["bold_number"],
-                ],
-                verbose=False,
-            )
-
-            # append to endlogs
-            endlogs.append(endlog)
-
-            # run caret_command
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                bsource,
-                "caret_command -file-convert -vc %s %s"
-                % (f["bold1"].replace("img", "ifh"), bsource),
-                "converting %s to nifti" % (f["bold1"]),
-                overwrite=overwrite,
-                remove=options["log"] == "remove",
-                thread=sinfo["id"],
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["logtag"],
-                    "B%d" % boldinfo["bold_number"],
-                ],
-                verbose=False,
-            )
-
-            # append to endlogs
-            endlogs.append(endlog)
-
-        # --- run BET
-        if os.path.exists(bbtarget) and not overwrite:
-            log.detail(f"bet on {os.path.basename(bsource)} already run")
-            report["bolddone"] += 1
-        else:
-            # run BET
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                bbtarget,
-                "bet %s %s %s" % (bsource, bbtarget, options["betboldmask"]),
-                "    ... running BET on %s with options %s"
-                % (os.path.basename(bsource), options["betboldmask"]),
-                overwrite=overwrite,
-                remove=options["log"] == "remove",
-                thread=sinfo["id"],
-                task="bet",
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["logtag"],
-                    "B%d" % boldinfo["bold_number"],
-                ],
-                verbose=False,
-            )
-            report["boldok"] += 1
-
-            # append to endlogs
-            endlogs.append(endlog)
-
-        if options["image_target"] == "4dfp":
-            # --- convert nifti to 4dfp
-            # run gunzip
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                bbtarget,
-                "gunzip -f %s.gz" % (bbtarget),
-                "    ... gunzipping %s.gz" % (os.path.basename(bbtarget)),
-                overwrite=overwrite,
-                remove=options["log"] == "remove",
-                thread=sinfo["id"],
-                task="gunzip",
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["logtag"],
-                    "B%d" % boldinfo["bold_number"],
-                ],
-                verbose=False,
-            )
-
-            # append to endlogs
-            endlogs.append(endlog)
-
-            # run gunzip
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                bmtarget,
-                "gunzip -f %s.gz" % (bmtarget),
-                "    ... gunzipping %s.gz" % (os.path.basename(bmtarget)),
-                overwrite=overwrite,
-                remove=options["log"] == "remove",
-                thread=sinfo["id"],
-                task="gunzip",
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["logtag"],
-                    "B%d" % boldinfo["bold_number"],
-                ],
-                verbose=False,
-            )
-
-            # append to endlogs
-            endlogs.append(endlog)
-
-            # run g_FlipFormat
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                f["bold1_brain"],
-                "g_FlipFormat %s %s"
-                % (bbtarget, f["bold1_brain"].replace(".img", ".ifh")),
-                "    ... converting %s to 4dfp" % (f["bold1_brain_nifti"]),
-                overwrite=overwrite,
-                remove=options["log"] == "remove",
-                thread=sinfo["id"],
-                task="FlipFormat",
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["logtag"],
-                    "B%d" % boldinfo["bold_number"],
-                ],
-                verbose=False,
-            )
-
-            # append to endlogs
-            endlogs.append(endlog)
-
-            # run g_FlipFormat
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                f["bold1_brain_mask"],
-                "g_FlipFormat %s %s"
-                % (bmtarget, f["bold1_brain_mask"].replace(".img", ".ifh")),
-                "    ... converting %s to 4dfp" % (f["bold1_brain_mask_nifti"]),
-                overwrite=overwrite,
-                remove=options["log"] == "remove",
-                thread=sinfo["id"],
-                task="FlipFormat",
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["logtag"],
-                    "B%d" % boldinfo["bold_number"],
-                ],
-                verbose=False,
-            )
-
-            # append to endlogs
-            endlogs.append(endlog)
-
-        else:
-            # --- link a template
-            if options["run"] != "run":
-                # `fl.lock` writes a .lock file beside the template, so a dry
-                # run cannot take the lock, let alone make the link it guards
-                log.detail(f"test, not linked: {os.path.basename(f['bold1_brain'])} as the bold template")
-            else:
-                # lock
-                fl.lock(templatefile)
-
-                # create link
-                if not os.path.exists(templatefile):
-                    # r += '\n ... link %s to %s' % (f['bold1_brain'], f['bold_template'])
-                    gc.link_or_copy(f["bold1_brain"], f["bold_template"])
-
-                # unlock
-                fl.unlock(templatefile)
-
-    except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-        log.raw(str(errormessage))
-
-        # unlock tempalte file if it crashed there
-        fl.unlock(templatefile)
-
-        report["boldfail"] += 1
-    except Exception as e:
-        # unlock template file if it crashed there
-        if templatefile is not None and os.path.exists(templatefile):
-            fl.unlock(templatefile)
-
-        report["boldfail"] += 1
-        log.error(f"Unknown error occured: \n...................................\n{e}\n{traceback.format_exc()}...................................\n")
-        time.sleep(1)
-
-    # merge into final comlog
-    log_prefix = "done"
-
-    # final_ log storage
-    final_log = ""
-
-    # if remove option is set do nothinng
-    remove = options["log"] == "remove"
-    if not remove and endlogs:
-        for el in endlogs:
-            # a dry run ran nothing, so every entry is the None `_run_external`
-            # returns in place of a comlog path
-            if el is not None and os.path.exists(el):
-                # did the command error out?
-                el_log = os.path.basename(el)
-                if "error" in el_log:
-                    log_prefix = "error"
-
-                # read log
-                with open(el, "r") as f:
-                    log_content = f.read()
-
-                # concatenate
-                final_log = final_log + log_content + "\n\n"
-
-                # delete the partial log
-                _remove(log, options, el)
-
-    # fails?
-    if report["boldfail"] > 0:
-        log_prefix = "error"
-    elif not overwrite and final_log == "":
-        final_log = "Previous results present, overwrite set to no.\n\n"
-        final_log = final_log + f"---> Successful completion of task at {datetime.now()}"
-
-    # a dry run ran nothing, so there is no output to merge and no comlog to
-    # leave behind -- the same rule `pc.combined_comlog` follows
-    if options["run"] != "run":
-        return {"r": log.text, "report": report}
-
-    # print to log file
-    logstamp = datetime.now().strftime("%Y-%m-%d_%H.%M.%S.%f")
-    logname = "%s_create_bold_brain_masks_B%s_%s_%s.log" % (
-        log_prefix,
-        boldinfo["bold_number"],
-        sinfo["id"],
-        logstamp,
-    )
-
-    # setup log folder
-    logfolder = options["comlogs"]
-    logfolders = []
-    if type(logfolder) in [list, set, tuple]:
-        logfolders = list(logfolder)
-        logfolder = logfolders.pop(0)
-
-    if not os.path.exists(logfolder):
-        try:
-            os.makedirs(logfolder)
-        except Exception:
-            raise pc.ExternalFailed(
-                "\n\nERROR: Could not create folder for logfile [%s]!" % (logfolder)
-            )
-
-    # print to file and close
-    logfile = os.path.join(logfolder, logname)
-    lf = open(logfile, "a")
-    lf.write(final_log)
-    lf.close()
-
-    return {"r": log.text, "report": report}
-
-
-def compute_bold_stats(sinfo, options, overwrite=False, thread=0):
-    """
-    ``compute_bold_stats [... processing options]``
-
-    Process specified BOLD files and save images/function/movement files.
-
-    ..  qx_command:
-        type: processing.session
-
-    Parameters:
-        --batchfile (str, default ''):
-            The batch.txt file with all the session information.
-
-        --sessionsfolder (str, default '.'):
-            The path to the study/sessions folder, where the imaging data is
-            supposed to go.
-
-        --parsessions (int, default 1):
-            How many sessions to run in parallel.
-
-        --parelements (int, default 1):
-            How many elements (e.g. bolds) to run in parallel.
-
-        --overwrite (str, default 'no'):
-            Whether to overwrite existing data (yes) or not (no). Note that
-            previous data is deleted before the run, so in the case of a failed
-            command run, previous results are lost.
-
-        --bolds (str, default 'rest'):
-            Which bold images (as they are specified in the batch.txt file) to
-            copy over. It can be a single type (e.g. 'task'), a pipe separated
-            list (e.g. 'WM|Control|rest') or 'all' to copy all.
-
-        --boldname (str, default 'bold'):
-            The default name of the bold files in the images folder.
-
-        --nifti_tail (str, default ''):
-            The tail of NIfTI volume images to use.
-
-        --bold_variant (str, default ''):
-            Optional variant of bold preprocessing. If specified, the BOLD
-            images in `images/functional<bold_variant>` will be processed.
-
-        --img_suffix (str, default ''):
-            Specifies a suffix for 'images' folder to enable support for
-            multiple parallel workflows. Empty if not used.
-
-        --logfolder (str, default ''):
-            The path to the folder where logs are to be stored,
-            if other than default.
-
-        --mov_radius (int, default 50):
-            Estimated head radius (in mm) for computing frame displacement
-            statistics.
-
-        --mov_fd (float, default 0.5):
-            Frame displacement threshold (in mm) to use for identifying bad
-            frames.
-
-        --mov_dvars (float, default 3.0):
-            The (mean normalized) dvars threshold to use for identifying bad
-            frames.
-
-        --mov_dvarsme (float, default 1.5):
-            The (median normalized) dvarsm threshold to use for identifying bad
-            frames.
-
-        --mov_after (int, default 0):
-            How many frames after each frame identified as bad to also exclude
-            from further processing and analysis.
-
-        --mov_before (int, default 0):
-            How many frames before each frame identified as bad to also exclude
-            from further processing and analysis.
-
-        --mov_bad (str, default 'udvarsme'):
-            Which criteria to use for identification of bad frames (mov, dvars,
-            dvarsme, idvars, uvars, idvarsme, udvarsme). See movement scrubbing
-            documentation for further information.
-            Criteria for identification of bad frames can be one out of:
-
-            - 'mov'      ... Frame displacement threshold (fdt) is exceeded.
-            - 'dvars'    ... Image intensity normalized root mean squared error
-              (RMSE) threshold (dvarsmt) is exceeded.
-            - 'dvarsme'  ... Median normalised RMSE (dvarsmet) threshold is
-              exceeded.
-            - 'idvars'   ... Both fdt and dvarsmt are exceeded (i for
-              intersection).
-            - 'uvars'    ... Either fdt or dvarsmt are exceeded (u for union).
-            - 'idvarsme' ... Both fdt and dvarsmet are exceeded.
-            - 'udvarsme' ... Either fdt or udvarsmet are exceeded.
-
-            For more detailed description please see wiki entry on Movement
-            scrubbing.
-
-    Notes:
-        The parameters listed in `Other parameters` can be specified in command
-        call or session.txt file.
-
-        The compute_bold_stats function processes each of the specified BOLD
-        files and saves three files in the images/functional/movement folder:
-
-        bold[N].bstats:
-            bold[N]<nifti_tail>.bstats includes for each frame of the BOLD image
-            the following computed statistics:
-
-            - n
-                Number of brain voxels.
-            - m
-                Mean signal intensity across all brain voxels.
-            - var
-                Signal variance across all brain voxels.
-            - sd
-                Signal standard variation across all brain voxels.
-            - dvars
-                RMDS measure of signal intensity difference between this and the
-                preceeding frame.
-            - dvarsm
-                Mean normalized dvars measure.
-            - dvarsme
-                Median normalized dvarsm measure.
-            - fd
-                Frame displacement.
-
-            There are three additional lines at the end of the file listing
-            maximum, mean and standard deviation of values across all timepoints
-            / volumes.
-
-        bold[N].scrub:
-            bold[N]<nifti_tail>.scrub includes for each frame the information on
-            whether the frame should be excluded (1) or not (0) based on the
-            following criteria (note below the relevant settings that specify
-            thresholds etc.):
-
-            - mov
-                Is frame displacement higher from the specified threshold?
-            - dvars
-                Is mean normalized dvars (dvarsm) higher than the specified
-                threshold?
-            - dvarsme
-                Is the median normalized dvarsm higher than the specified
-                threshold?
-            - idvars
-                Are both frame displacement as well as dvarsm measures above
-                threshold (intersection of fd and dvarsm).
-            - idvarsme
-                Are both frame displacement as well as dvarsme measures above
-                threshold (intersection of fd and dvarsme).
-            - udvars
-                Are either frame displacement or dvarsm measures above threshold
-                (union of fs and dvarsm).
-            - udvarsme
-                Are either frame displacement or dvarsme measures above
-                threshold (union of fs and dvarsme).
-
-            The last column of the file is a 'use' column, which specifies,
-            based on the criteria provided, whether the frame should be used in
-            further preprocessing and analysis or not.
-
-            There is an additional #sum line at the end of the file, listing how
-            many frames are marked as bad using each criteria.
-
-        bold[N].use:
-            bold[N]<nifti_tail>.use file lists for each frame of the relevant
-            BOLD image, whether it is to be used (1) or not (0).
-
-        When 'cifti' is the specified image target, the related nifti volume
-        files will be processed as only they provide all the information for
-        computing the relevant parameters.
-
-        Dependencies:
-            The command runs the general_compute_bold_stats.m Matlab function
-            for computation of parameters. It also expects that both bold images
-            and the related movement correction parameter files are present in
-            the expected locations.
-
-    Examples:
-        Using the defaults::
-
-            qunex compute_bold_stats \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --bolds=all
-
-        Specifying additional parameters for identification of bad frames::
-
-            qunex compute_bold_stats \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --bolds=all \\
-                --mov_fd=0.9 \\
-                --mov_dvarsme=1.6 \\
-                --mov_before=1 \\
-                --mov_after=2
-    """
-    log = ReportLog()
-
-    report = {
-        "bolddone": 0,
-        "boldok": 0,
-        "boldfail": 0,
-        "boldmissing": 0,
-        "boldskipped": 0,
-    }
-
-    log.rule()
-    log.info(f"Session id: {sinfo['id']} \n[started on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}]")
-    log.action("Computing", "BOLD image statistics ...", options["run"], level="info")
-    log.blank()
-    log.info(BOLD_STATS_PURPOSE)
-
-    pc.do_options_check(options, sinfo, "compute_bold_stats")
-    d = pc.get_session_folders(sinfo, options)
-
-    if overwrite:
-        ostatus = "will"
-    else:
-        ostatus = "will not"
-
-    log.step("Using parameters for computing scrubbing information")
-    for name in BOLD_STATS_PARAMETERS:
-        log.detail(f"--{name}: {options[name]}")
-
-    log.step(f"Working on BOLD images in {d['s_bold']}")
-    log.detail(f"images{options['img_suffix']}/functional{options['bold_variant']} will be processed")
-    log.detail(f"the resulting files will be in {d['s_bold_mov']}")
-    log.detail(f"{', '.join(options['bolds'].split('|'))} BOLD files will be processed (see --bolds)")
-    log.detail(f"existing statistics {ostatus} be overwritten (see --overwrite)")
-
-    bolds, bskip, report["boldskipped"] = pc.use_or_skip_bold(sinfo, options, _log=log)
-
-    parelements = options["parelements"]
-    log.info(f"Processing {parelements} BOLDs in parallel")
-
-    if parelements == 1:  # serial execution
-        for b in bolds:
-            # process
-            result = execute_compute_bold_stats(sinfo, options, overwrite, b)
-
-            # merge r
-            log.raw(result["r"])
-
-            # merge report
-            temp_report = result["report"]
-            report["bolddone"] += temp_report["bolddone"]
-            report["boldok"] += temp_report["boldok"]
-            report["boldfail"] += temp_report["boldfail"]
-            report["boldmissing"] += temp_report["boldmissing"]
-    else:  # parallel execution
-        # create a multiprocessing Pool
-        process_pool_executor = ProcessPoolExecutor(parelements)
-        # process
-        f = partial(execute_compute_bold_stats, sinfo, options, overwrite)
-        results = process_pool_executor.map(f, bolds)
-
-        # merge r and report
-        for result in results:
-            log.raw(result["r"])
-            temp_report = result["report"]
-            report["bolddone"] += temp_report["bolddone"]
-            report["boldok"] += temp_report["boldok"]
-            report["boldfail"] += temp_report["boldfail"]
-            report["boldmissing"] += temp_report["boldmissing"]
-
-    log.blank()
-    log.info(f"Bold statistics computation completed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-    log.rule()
-    rstatus = (
-        "BOLDS done: %(bolddone)2d, missing data: %(boldmissing)2d, failed: %(boldfail)2d, processed: %(boldok)2d, skipped: %(boldskipped)2d"
-        % (report)
-    )
-
-    # print r
-    return log.result(rstatus, report["boldmissing"] + report["boldfail"], sinfo["id"])
-
-
-def execute_compute_bold_stats(sinfo, options, overwrite, boldinfo):
-
-    # prepare return variables
-    log = ReportLog()
-    report = {"bolddone": 0, "boldok": 0, "boldfail": 0, "boldmissing": 0}
-
-    log.step("Working on " + boldinfo["name"] + " ...")
-
-    try:
-        # --- filenames
-
-        f = pc.get_file_names(sinfo, options)
-        f.update(pc.get_bold_file_names(sinfo, boldinfo["name"], options))
-        d = pc.get_session_folders(sinfo, options)
-
-        # --- check for data availability
-
-        log.detail("checking for data")
-        status = True
-
-        # --- movement
-        status = pc.check_for_file(f["bold_mov"],
-            f"movement data present [{os.path.basename(f['bold_mov'])}]",
-            f"movement data missing [{os.path.basename(f['bold_mov'])}]",
-            status=status,
-            _log=log,
-        )
-
-        # --- bold
-        status = pc.check_for_file(f["bold_vol"],
-            f"bold data present [{os.path.basename(f['bold_vol'])}]",
-            f"bold data missing [{os.path.basename(f['bold_vol'])}]",
-            status=status,
-            _log=log,
-        )
-
-        # --- check
-        if not status:
-            log.error("Files missing, skipping this bold run!")
-            report["boldmissing"] += 1
-            return {"r": log.text, "report": report}
-
-        # --- running the stats
-
-        scrub = (
-            "radius:%d|fdt:%.2f|dvarsmt:%.2f|dvarsmet:%.2f|after:%d|before:%d|reject:%s"
-            % (
-                options["mov_radius"],
-                options["mov_fd"],
-                options["mov_dvars"],
-                options["mov_dvarsme"],
-                options["mov_after"],
-                options["mov_before"],
-                options["mov_bad"],
-            )
-        )
-        comm = (
-            "%s \"try general_compute_bold_stats('%s', '', '%s', 'same', '%s', true); catch ME, general_report_crash(ME); exit(1), end; exit\""
-            % (mcommand, f["bold_vol"], d["s_bold_mov"], scrub)
-        )
-        if options["print_command"] == "yes":
-            log.pipeline_command(comm, title="Running:")
-        runit = True
-        if os.path.exists(f["bold_stats"]) and not overwrite:
-            report["bolddone"] += 1
-            runit = False
-        endlog, status, failed = _run_external(
-            log,
-            options,
-            f["bold_stats"],
-            comm,
-            "... running matlab general_compute_bold_stats on %s" % (f["bold_vol"]),
-            overwrite=overwrite,
-            thread=sinfo["id"],
-            remove=options["log"] == "remove",
-            task=options["command_ran"],
-            logfolder=options["comlogs"],
-            logtags=[
-                options["bold_variant"],
-                options["logtag"],
-                "B%d" % boldinfo["bold_number"],
-            ],
-            shell=True,
-        )
-        status = pc.check_for_file(
-            f["bold_stats"],
-            bad=f"Matlab/Octave has failed preprocessing BOLD using command: {comm}",
-            bad_level="error",
-            _log=log,
-        )
-
-        if status and runit:
-            report["boldok"] += 1
-        elif runit:
-            report["boldfail"] += 1
-
-    except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-        log.raw(str(errormessage))
-        report["boldfail"] += 1
-    except Exception:
-        log.error(f"Unknown error occured: \n...................................\n{traceback.format_exc()}...................................\n")
-        report["boldfail"] += 1
-
-    return {"r": log.text, "report": report}
-
-
-def create_stats_report(sinfo, options, overwrite=False, thread=0):
-    """
-    ``create_stats_report [... processing options]``
-
-    Process movement correction parameters and computed BOLD statistics to
-    create per session plots and fidl snippets and group reports.
-
-    ..  qx_command:
-        type: processing.session
-
-    Parameters:
-        --batchfile (str, default ''):
-            The batch.txt file with all the session information.
-
-        --sessionsfolder (str, default '.'):
-            The path to the study/sessions folder, where the imaging data is
-            supposed to go.
-
-        --parsessions (int, default 1):
-            How many sessions to run in parallel.
-
-        --overwrite (str, default 'no'):
-            Whether to overwrite existing data (yes) or not (no). Note that
-            previous data is deleted before the run, so in the case of a failed
-            command run, previous results are lost.
-
-        --bolds (str, default 'rest'):
-            Which bold images (as they are specified in the batch.txt file) to
-            copy over. It can be a single type (e.g. 'task'), a pipe separated
-            list (e.g. 'WM|Control|rest') or 'all' to copy all.
-
-        --boldname (str, default 'bold'):
-            The default name of the bold files in the images folder.
-
-        --nifti_tail (str, default ''):
-            The tail of NIfTI volume images to use.
-
-        --bold_variant (str, default ''):
-            Optional variant of bold preprocessing. If specified, the BOLD
-            images in `images/functional<bold_variant>` will be processed.
-
-        --img_suffix (str, default ''):
-            Specifies a suffix for 'images' folder to enable support for
-            multiple parallel workflows. Empty if not used.
-
-        --logfolder (str, default ''):
-            The path to the folder where logs are to be stored,
-            if other than default.
-
-        --mov_radius (int, default 50):
-            Estimated head radius (in mm) for computing frame displacement
-            statistics.
-
-        --mov_fd (float, default 0.5):
-            Frame displacement threshold (in mm) to use for identifying bad
-            frames.
-
-        --mov_dvars (float, default 3.0):
-            The (mean normalized) dvars threshold to use for identifying bad
-            frames. Later referred to as dvarsmt.
-
-        --mov_dvarsme (float, default 1.5):
-            The (median normalized) dvarsm threshold to use for identifying bad
-            frames. Later referred to as dvarsmet.
-
-        --mov_after (int, default 0):
-            How many frames after each frame identified as bad to also exclude
-            from further processing and analysis.
-
-        --mov_before (int, default 0):
-            How many frames before each frame identified as bad to also exclude
-            from further processing and analysis.
-
-        --tr (float, default 2.5):
-            TR of the BOLD files.
-
-        --mov_pref (str, default ''):
-            The prefix to be used for the figure plot files.
-
-        --mov_plot (str, default 'mov_report'):
-            The base name of the plot files. If set to empty no plots are
-            generated.
-
-        --mov_mreport (str, default 'movement_report.txt'):
-            The name of the group movement report file. If set to an empty
-            string, no file is generated.
-
-        --mov_sreport (str, default 'movement_scrubbing_report.txt'):
-            The name of the group scrubbing report file. If set to an empty
-            string, no file is generated.
-
-        --mov_preport (str, default 'movement_report_post.txt'):
-            The name of group report file with stats computed with frames
-            identified as bad exluded from analysis. If set to an empty
-            string, no file is generated.
-
-        --mov_post (str, default 'udvarsme'):
-            The criterium for identification of bad frames that is used when
-            generating a post scrubbing statistics group report. The value names
-            a column of the `.scrub` file that `compute_bold_stats` wrote:
-
-            - 'mov'       ... Frame displacement threshold (fdt) is exceeded.
-            - 'dvars'     ... Image intensity normalized root mean squared error
-              (RMSE) threshold (dvarsmt) is exceeded.
-            - 'dvarsme'   ... Median normalised RMSE threshold (dvarsmet) is
-              exceeded.
-            - 'idvars'    ... Both fdt and dvarsmt are exceeded (i for
-              intersection).
-            - 'idvarsme'  ... Both fdt and dvarsmet are exceeded.
-            - 'udvars'    ... Either fdt or dvarsmt are exceeded (u for union).
-            - 'udvarsme'  ... Either fdt or udvarsmet are exceeded (default).
-            - 'none'.
-
-            For more detailed description please see wiki entry on Movement
-            scrubbing.
-
-        --mov_fidl (str, default 'udvarsme'):
-            Whether to create fidl file snippets with listed bad frames, and
-            what criterium to use for the definition of bad frames. The value
-            names a column of the `.scrub` file that `compute_bold_stats` wrote:
-
-            - 'mov'       ... Frame displacement threshold (fdt) is exceeded.
-            - 'dvars'     ... Image intensity normalized root mean squared error
-              (RMSE) threshold (dvarsmt) is exceeded.
-            - 'dvarsme'   ... Median normalised RMSE threshold  (dvarsmet) is
-              exceeded.
-            - 'idvars'    ... Both fdt and dvarsmt are exceeded (i for
-              intersection).
-            - 'idvarsme'  ... Both fdt and dvarsmet are exceeded.
-            - 'udvars'    ... Either fdt or dvarsmt are exceeded (u for union).
-            - 'udvarsme'  ... Either fdt or udvarsmet are exceeded (default).
-            - 'none'      ... Set to 'none' to not generate them.
-
-            For more detailed description please see wiki entry on Movement
-            scrubbing.
-
-        --mov_pdf (str, default 'movement_plots'):
-            The name of the folder in sessions/QC/movement in which to copy
-            the individuals' movement plots.
-
-    Notes:
-        Use:
-            create_stats_report processes movement correction parameters and
-            computed BOLD statistics to create per session plots and fidl
-            snippets and group reports.
-
-            For each session it saves into images/functional/movement:
-
-            --`bold<nifti_tail>_<mov_plot>_cor.pdf`
-                A plot of movement correction parameters for each of the BOLD
-                files.
-            --`bold<nifti_tail>_<mov_plot>_dvars.pdf`
-                A plot of frame displacement and dvarsm statistics with frames
-                that are identified as bad marked in blue.
-            --`bold<nifti_tail>_<mov_plot>_dvarsme.pdf`
-                A plot of frame displacement and dvarsme statistics with frames
-                that are identified as bad marked in blue.
-            --`bold[N]<nifti_tail>_scrub.fidl`
-                A fidl filesnippet that lists, which frames are to be excluded
-                from the analysis.
-
-            For the group level it creates three report files that are stored in
-            the <sessionsfolder>/QC/movement folder. These files are:
-
-            - ``<mov_mreport>`` (bold<nifti_tail>_movement_report.txt by default)
-                This file lists for each session and bold file mean, sd, range,
-                max, min, median, and squared mean divided by max statistics for
-                each of the 6 movement correction parameters. It also prints
-                mean, median, maximum, and standard deviation of frame
-                displacement statistics. The purpose of this file is to enable
-                easy session and group level analysis of movement in the scanner.
-
-            - ``<mov_preport>`` (bold<nifti_tail>_movement_report_post.txt by default)
-                This file has the same structure and information as the above,
-                with frames marked as bad excluded from the statistics
-                computation. This enables session and group level assessment of
-                the effects of scrubbing.
-
-            - ``<mov_sreport>`` (bold<nifti_tail>_movement_scrubbing_report.txt by default)
-                This file lists for each BOLD of each session the number and the
-                percentage of frames that would be marked as bad and excluded
-                from the analyses when a specific exclusion criteria would be
-                used. Again, the file supports session and group level analysis
-                of movement scrubbing.
-
-            Extra notes and dependencies:
-                The statistics are computed and the plots drawn in Python. The
-                command requires that movement correction parameters files and
-                bold statistics data files (results of the compute_bold_stats
-                command) are present in the expected locations.
-
-                Session statistics are appended to the group level report files
-                as they are being computed. To avoid messy group level files, it
-                is recommended to run the command with parsessions set to 1
-                (example 1), to enforce sequential processing and adding of
-                information to group level statistics files. Another option is
-                to run the processing in two steps. The first step with multiple
-                parsessions to speed up generation of session level maps
-                (example 2), and then the second step with a single core,
-                omitting the slow generation of session specific plots.
-
-    Examples:
-        ::
-
-            qunex create_stats_report \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --bolds=all \\
-                --parsessions=1
-
-        ::
-
-            qunex create_stats_report \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --bolds=all \\
-                --parsessions=10
-
-        ::
-
-            qunex create_stats_report \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --bolds=all \\
-                --nifti_tail=_hp2000_clean \\
-                --parsessions=1 \\
-                --mov_plot=""
-    """
-    log = ReportLog()
-
-    preport = {
-        "plotdone": "done",
-        "boldok": 0,
-        "procok": "ok",
-        "boldmissing": 0,
-        "boldskipped": 0,
-    }
-
-    try:
-        log.rule()
-        log.info(f"Session id: {sinfo['id']} \n[started on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}]")
-        log.action("Creating", "BOLD movement and statistics report ...", options["run"], level="info")
-        log.blank()
-        log.info(STATS_REPORT_PURPOSE)
-
-        pc.do_options_check(options, sinfo, "create_stats_report")
-        d = pc.get_session_folders(sinfo, options)
-
-        if overwrite:
-            ostatus = "will"
-        else:
-            ostatus = "will not"
-
-        log.step("Using parameters")
-        for name in STATS_REPORT_PARAMETERS:
-            log.detail(f"--{name}: {options[name]}")
-
-        log.step(f"Working on BOLD information in {d['s_bold_mov']}")
-        log.detail(f"images{options['img_suffix']}/functional{options['bold_variant']} will be processed")
-        log.detail("the resulting plots will be saved there")
-        log.detail(f"{', '.join(options['bolds'].split('|'))} BOLD files will be processed (see --bolds)")
-        log.detail(f"existing results {ostatus} be overwritten (see --overwrite)")
-
-        procbolds = []
-        d = pc.get_session_folders(sinfo, options)
-
-        # --- check for data
-
-        if options["mov_plot"] != "":
-            if (
-                os.path.exists(
-                    os.path.join(
-                        d["s_bold_mov"],
-                        options["mov_pref"] + options["mov_plot"] + "_cor.pdf",
-                    )
-                )
-                and not overwrite
-            ):
-                log.detail("Movement plots already exists! Please use option --overwrite=yes to redo them!")
-                preport["plotdone"] = "old"
-                plot = ""
-            else:
-                plot = options["mov_plot"]
-                preport["plotdone"] = "new"
-        else:
-            plot = ""
-            preport["plotdone"] = "none"
-
-        log.step(f"Checking for data in {d['s_bold_mov']}")
-
-        bolds, bskip, preport["boldskipped"] = pc.use_or_skip_bold(sinfo, options, _log=log)
-
-        for boldinfo in bolds:
-            log.step("Working on " + boldinfo["name"] + " ...")
-
-            try:
-                # --- filenames
-                f = pc.get_file_names(sinfo, options)
-                f.update(pc.get_bold_file_names(sinfo, boldinfo["name"], options))
-
-                # --- check for data availability
-
-                status = True
-
-                if os.path.exists(d["s_bold_mov"]):
-                    # --- movement
-                    status = pc.check_for_file(f["bold_mov"],
-                        f"movement data present [{os.path.basename(f['bold_mov'])}]",
-                        f"movement data missing [{os.path.basename(f['bold_mov'])}]",
-                        status=status,
-                        _log=log,
-                    )
-                    status = pc.check_for_file(f["bold_stats"],
-                        f"stats data present [{os.path.basename(f['bold_stats'])}]",
-                        f"stats data missing [{os.path.basename(f['bold_stats'])}]",
-                        status=status,
-                        _log=log,
-                    )
-                    status = pc.check_for_file(f["bold_scrub"],
-                        f"scrub data present [{os.path.basename(f['bold_scrub'])}]",
-                        f"scrub data missing [{os.path.basename(f['bold_scrub'])}]",
-                        status=status,
-                        _log=log,
-                    )
-                else:
-                    log.detail("folder does not exist!")
-                    status = False
-
-                # --- check
-                if status:
-                    procbolds.append(boldinfo["bold_number"])
-                    preport["boldok"] += 1
-                else:
-                    log.error("Files missing, skipping this bold run!")
-                    preport["boldmissing"] += 1
-
-            except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-                log.raw(str(errormessage))
-            except Exception:
-                log.error(f"Unknown error occured: \n...................................\n{traceback.format_exc()}...................................\n")
-
-        # build the reports
-
-        procbolds.sort()
-        procbolds = [str(e) for e in procbolds]
-
-        report = {}
-
-        for tf in ["mov_mreport", "mov_sreport", "mov_preport"]:
-            if options[tf] != "":
-                tmpf = os.path.join(
-                    d["qc_mov"],
-                    options["boldname"] + options["nifti_tail"] + "_" + options[tf],
-                )
-                report[tf] = tmpf
-                if os.path.exists(tmpf) and thread == 1:
-                    _remove(log, options, tmpf)
-            else:
-                report[tf] = ""
-
-        runs = [options["boldname"] + e + options["nifti_tail"] for e in procbolds]
-
-        # one line for both arms: `action` is what spells it "Test running ..."
-        # under `--test`, so the run mode is the log's business and not a branch
-        log.action(
-            "Running",
-            f"movement and statistics reporting for {', '.join(runs)}",
-            options["run"],
-        )
-
-        if options["run"] != "run":
-            for tf in ["mov_mreport", "mov_sreport", "mov_preport"]:
-                if report[tf] != "":
-                    log.detail(f"test, not written: {report[tf]}")
-            if options["mov_fidl"] != "none":
-                for run in runs:
-                    log.detail(f"test, not written: {os.path.join(d['s_bold_mov'], run + '_scrub.fidl')}")
-            preport["procok"] = "test"
-        else:
-            # failure arrives as an exception now, caught by the handlers below
-            ms.report_movement_statistics(
-                d["s_bold_mov"], procbolds, sinfo["id"], report, options, plot, _log=log
-            )
-            preport["procok"] = "ok"
-
-        if options["mov_plot"] != "" and options["mov_pdf"] != "no":
-            for sf in ["cor", "dvars", "dvarsme"]:
-                tfolder = os.path.join(d["qc_mov"], options["mov_pdf"], sf)
-                if not os.path.exists(tfolder):
-                    if options["run"] != "run":
-                        log.detail(f"test, not created: {tfolder}")
-                    else:
-                        os.makedirs(tfolder)
-
-                froot = "%s%s_%s%s_%s.pdf" % (
-                    options["boldname"],
-                    options["nifti_tail"],
-                    options["mov_pref"],
-                    options["mov_plot"],
-                    sf,
-                )
-                if os.path.exists(
-                    os.path.join(tfolder, "%s-%s" % (sinfo["id"], froot))
-                ):
-                    _remove(
-                        log,
-                        options,
-                        os.path.join(tfolder, "%s-%s" % (sinfo["id"], froot)),
-                    )
-                _link_or_copy(
-                    log,
-                    options,
-                    os.path.join(d["s_bold_mov"], froot),
-                    os.path.join(tfolder, "%s-%s" % (sinfo["id"], froot)),
-                )
-                if options["run"] == "run":
-                    log.detail(f"copying {os.path.join(d['s_bold_mov'], froot)} to {os.path.join(tfolder, '%s-%s' % (sinfo['id'], froot))}")
-
-        if (
-            options["mov_fidl"] in ms.CRITERIA
-            and options["event_file"] != ""
-            and options["bolds"] != ""
-        ):
-            concf = os.path.join(d["s_bold_concs"], options["bolds"] + ".conc")
-            fidlf = os.path.join(d["s_bold_events"], options["event_file"] + ".fidl")
-            ipatt = "_%s_scrub.fidl" % (options["mov_fidl"])
-
-            if options["run"] != "run":
-                log.detail(
-                    f'test, not written: {fidlf.replace(".fidl", ipatt)}'
-                )
-            elif os.path.exists(concf) and os.path.exists(fidlf):
-                try:
-                    gm.meltmovfidl(concf, ipatt, fidlf, fidlf.replace(".fidl", ipatt))
-                except Exception:
-                    log.warning(
-                        f"Failed to create a melted fidl file! ({sinfo['id']})"
-                    )
-                    raise
-            else:
-                log.warning("Files missing, failed to create a melted fidl file!")
-
-    except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-        log.raw(str(errormessage))
-        log.info(f"BOLD statistics and movement report failed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-        log.rule()
-        preport["procok"] = "failed"
-    except Exception:
-        log.info(f"BOLD statistics and movement report failed with and unknown error: \n...................................\n{traceback.format_exc()}...................................\n")
-        preport["procok"] = "failed"
-
-    if preport["procok"] == "ok":
-        log.blank()
-        log.info(f"BOLD statistics and movement report completed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-        log.rule()
-
-    rstatus = (
-        "BOLDs ok: %(boldok)2d, missing data: %(boldmissing)2d, processing: %(procok)s, skipped: %(boldskipped)s"
-        % (preport)
-    )
-    if preport["procok"] == "ok":
-        rstatus += ", plots: %(plotdone)s" % (preport)
-
-    return log.result(
-        rstatus,
-        preport["boldmissing"] + (preport["procok"] == "failed"),
-        sinfo["id"],
-    )
-
-
-# -> @register_command(
-#        description="Extract nuisance signal from BOLD images.",
-#         type="processing.session")
-def extract_nuisance_signal(sinfo, options, overwrite=False, thread=0):
-    """
-    ``extract_nuisance_signal [... processing options]``
-
-    Extract nuisance signal from volume BOLD files.
-
-    ..  qx_command:
-        type: processing.session
-
-    Parameters:
-        --batchfile (str, default ''):
-            The batch.txt file with all the session information.
-
-        --sessionsfolder (str, default '.'):
-            The path to the study/sessions folder, where the imaging data is
-            supposed to go.
-
-        --parsessions (int, default 1):
-            How many sessions to run in parallel.
-
-        --parelements (int, default 1):
-            How many elements (e.g. bolds) to run in parallel.
-
-        --overwrite (str, default 'no'):
-            Whether to overwrite existing data (yes) or not (no). Note that
-            previous data is deleted before the run, so in the case of a failed
-            command run, previous results are lost.
-
-        --bolds (str, default 'rest'):
-            Which bold images (as they are specified in the batch.txt file) to
-            copy over. It can be a single type (e.g. 'task'), a pipe separated
-            list (e.g. 'WM|Control|rest') or 'all' to copy all.
-
-        --boldname (str, default 'bold'):
-            The default name of the bold files in the images folder.
-
-        --nifti_tail (str, default ''):
-            The tail of NIfTI volume images to use.
-
-        --bold_variant (str, default ''):
-            Optional variant of bold preprocessing. If specified, the BOLD
-            images in `images/functional<bold_variant>` will be processed.
-
-        --img_suffix (str, default ''):
-            Specifies a suffix for 'images' folder to enable support for
-            multiple parallel workflows. Empty if not used.
-
-        --logfolder (str, default ''):
-            The path to the folder where logs are to be stored,
-            if other than default.
-
-        --wbmask (str, default ''):
-            A path to an optional file that specifies which regions are to be
-            excluded from the whole-brain mask. It can be used in the case of
-            ROI analyses for which one does not want to include the ROI specific
-            signals in the global signal regression.
-
-        --nroi (str, default ''):
-            The path to additional nuisance regressors file. It can be either a
-            binary mask or a '.names' file that specifies the ROI to be used.
-            Based on other options, the ROI can be further masked by session
-            specific files or not masked at all (see Use section below).
-
-        --sessionroi (str, default ''):
-            A string specifying which session specific mask to use for further
-            masking the additional roi. The two options are 'wb' or 'aseg' for
-            whole brain mask or FreeSurfer aseg+aparc mask, respectively.
-
-        --shrinknsroi (str, default 'true'):
-            A string specifying whether to shrink ('true') the whole
-            brain and white matter masks or not ('false').
-
-    Output files:
-        The command generates the following files:
-
-        - `bold[N]<nifti_tail>.nuisance`
-              A text file that lists for each volume frame the information
-              on mean intensity across the ventricle, white matter and whole
-              brain voxels, and any additional nuisance ROI specified using
-              parameters. The file is stored in the
-              `images<img_suffix>/functional<bold_variant>/movement` folder.
-
-        - `bold[N]<nifti_tail>_nuisance.png`
-            A PNG image of axial slices of the first BOLD frame over which
-            the identified nuisance regions are overlayed. Ventricles in
-            green, white matter in red and the rest of the brain in blue.
-            The ventricle and white matter regions are defined based on
-            FreeSurfer segmentation. Each region is "trimmed" before use, so
-            that there is at least one voxel buffer between each nuisance
-            region mask. The image is stored in
-            images<img_suffix>/ROI/nuisance<bold_variant>`.
-
-        - `bold[N]<nifti_tail>_nuisance.<image format>`
-            An image file of the relevant image format that holds the same
-            information as the above PNG. It is a file of five volumes, the
-            first volume holds the first BOLD frame, the second the whole
-            brain mask, the third the ventricles mask and the fourth the
-            white matter mask. The fifth volume stores all three masks coded
-            as 1 (whole brain), 2 (ventricles), or 3 (white matter). The
-            image is stored in `images<img_suffix>/ROI/nuisance<bold_variant>`
-            folder.
-
-    Notes:
-        extract_nuisance_signal is used to extract nuisance signal from volume
-        BOLD files to be used in the latter steps of preprocessing,
-        specifically for regression of nuisance signals. By default, it
-        extracts nuisance signals from ventricles, white matter and whole
-        brain. Whole brain is defined as those parts of the brain that are not
-        ventricles or white matter, which results in whole brain to mostly
-        overlap with gray matter.
-
-        Using parameters listed below, it is also possible to specify
-        additional ROIs for which nuisance signal is to be extracted and/or ROI
-        that are to be excluded from the whole brain mask.
-
-        To exclude specific ROI from the whole brain mask, use the '--wbmask'
-        option. This should be a path to a file that specifies, which ROI are
-        to be excluded from the whole-brain mask. The reason for exclusion
-        might be when one does not want the signals from specific ROI to be
-        included in the global signal regression, thereby resolving some issues
-        taken as arguments against using global signal regression. The file can
-        be either a binary mask, or a '.names' file. In the latter case, it is
-        possible to additional mask the ROI to be excluded based on session
-        specific aseg+aparc image (see description of .names file format).
-
-        Another option is to include additional independent nuisance regions
-        that might or might not overlap with the existing masks. Two parameters
-        are used to specify this. The first is the '--nroi' parameter. This,
-        again, is a path to either a binary image or a '.names' file. In the
-        latter case, it is again possible to mask the additional ROI either by
-        the binary whole brain mask or the individuals aseg+aparc file. To
-        achieve this, set the additional '--sessionroi' parameter to 'wb' or
-        'aseg', respectively. If some additional ROI are to be excluded, even
-        though they fall outside the brain, then these are to be listed as
-        comma separated list of ROI names (that match the ROI names in the
-        .names file), separated from the path by a pipe ('|') symbol. For
-        instance if one also would like to include eyes and scull as two
-        additional nuiscance regions, one has to create a volume mask + a
-        .names file pair, and pass it as the '--nroi' parameter, e.g.::
-
-            --nroi="<path to ROI>/nroi.names|eyes,scull"
-
-        Extra notes and dependencies:
-            When 'cifti' is the specified image target, the related nifti volume
-            files will be processed as only they provide all the information for
-            computing the relevant parameters
-
-            The command runs the general_extract_nuisance.m Matlab function for
-            actual nuisance signal extraction. It expects that bold images,
-            whole brain masks, and aseg+aparc imags to be present in the
-            expected locations.
-
-    Examples:
-        ::
-
-            qunex extract_nuisance_signal \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --bolds=all \\
-                --parsessions=10
-    """
-    log = ReportLog()
-
-    report = {
-        "bolddone": 0,
-        "boldok": 0,
-        "boldfail": 0,
-        "boldmissing": 0,
-        "boldskipped": 0,
-    }
-
-    log.rule()
-    log.info(f"Session id: {sinfo['id']} \n[started on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}]")
-    log.action("Extracting", "BOLD nuisance signal ...", options["run"], level="info")
-    log.blank()
-    log.info(textwrap.fill(
-        NUISANCE_PURPOSE.format(
-            boldname=options["boldname"],
-            nifti_tail=options["nifti_tail"],
-            img_suffix=options["img_suffix"],
-        ),
-        PURPOSE_WIDTH,
-    ))
-
-    pc.do_options_check(options, sinfo, "extract_nuisance_signal")
-    d = pc.get_session_folders(sinfo, options)
-
-    if overwrite:
-        ostatus = "will"
-    else:
-        ostatus = "will not"
-
-    log.step("Using parameters when extracting nuisance signal")
-    for name in NUISANCE_PARAMETERS:
-        log.detail(f"--{name}: {options[name]}")
-
-    log.step(f"Working on BOLD images in {d['s_bold']}")
-    log.detail(f"images{options['img_suffix']}/functional{options['bold_variant']} will be processed")
-    log.detail(f"the resulting files will be in {d['s_bold_mov']}")
-    log.detail(f"{', '.join(options['bolds'].split('|'))} BOLD files will be processed (see --bolds)")
-    log.detail(f"existing nuisance files {ostatus} be overwritten (see --overwrite)")
-
-    bolds, bskip, report["boldskipped"] = pc.use_or_skip_bold(sinfo, options, _log=log)
-
-    parelements = options["parelements"]
-    log.info(f"Processing {parelements} BOLDs in parallel")
-
-    if parelements == 1:  # serial execution
-        for b in bolds:
-            # process
-            result = execute_extract_nuisance_signal(sinfo, options, overwrite, b)
-
-            # merge r
-            log.raw(result["r"])
-
-            # merge report
-            temp_report = result["report"]
-            report["bolddone"] += temp_report["bolddone"]
-            report["boldok"] += temp_report["boldok"]
-            report["boldfail"] += temp_report["boldfail"]
-            report["boldmissing"] += temp_report["boldmissing"]
-    else:  # parallel execution
-        # create a multiprocessing Pool
-        process_pool_executor = ProcessPoolExecutor(parelements)
-        # process
-        f = partial(execute_extract_nuisance_signal, sinfo, options, overwrite)
-        results = process_pool_executor.map(f, bolds)
-
-        # merge r and report
-        for result in results:
-            log.raw(result["r"])
-            temp_report = result["report"]
-            report["bolddone"] += temp_report["bolddone"]
-            report["boldok"] += temp_report["boldok"]
-            report["boldfail"] += temp_report["boldfail"]
-            report["boldmissing"] += temp_report["boldmissing"]
-
-    log.blank()
-    log.info(f"Bold nuisance signal extraction completed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-    log.rule()
-    rstatus = (
-        "BOLDS done: %(bolddone)2d, missing data: %(boldmissing)2d, failed: %(boldfail)2d, skipped: %(boldskipped)2d, processed: %(boldok)2d"
-        % (report)
-    )
-
-    return log.result(rstatus, report["boldmissing"] + report["boldfail"], sinfo["id"])
-
-
-def execute_extract_nuisance_signal(sinfo, options, overwrite, boldinfo):
-
-    # prepare return variables
-    log = ReportLog()
-    report = {"bolddone": 0, "boldok": 0, "boldfail": 0, "boldmissing": 0}
-
-    log.step("Working on " + boldinfo["name"] + " ...")
-
-    try:
-        # --- filenames
-        f = pc.get_file_names(sinfo, options)
-        f.update(pc.get_bold_file_names(sinfo, boldinfo["name"], options))
-        d = pc.get_session_folders(sinfo, options)
-
-        # --- check for data availability
-
-        log.detail("checking for data")
-        status = True
-
-        # --- bold mask
-        status = pc.check_for_file(f["bold1_brain_mask"],
-            "bold brain mask present",
-            f"bold brain mask missing [{f['bold1_brain_mask']}]",
-            status=status,
-            _log=log,
-        )
-
-        # --- aseg
-        astat = pc.check_for_file(f["fs_aseg_bold"],
-            "freesurfer aseg present",
-            f"freesurfer aseg missing [{f['fs_aseg_bold']}]",
-            status=True,
-            _log=log,
-        )
-        if not astat:
-            astat = pc.check_for_file(f["fs_aparc_bold"],
-                "freesurfer aparc present",
-                f"freesurfer aparc missing [{f['fs_aparc_bold']}]",
-                status=True,
-                _log=log,
-            )
-            segfile = f["fs_aparc_bold"]
-        else:
-            segfile = f["fs_aseg_bold"]
-
-        status = status and astat
-
-        # --- bold
-        status = pc.check_for_file(f["bold_vol"],
-            "bold data present",
-            f"bold data missing [{f['bold_vol']}]",
-            status=status,
-            _log=log,
-        )
-
-        # --- check
-        if not status:
-            log.error("Files missing, skipping this bold run!")
-            report["boldmissing"] += 1
-            return {"r": log.text, "report": report}
-
-        # --- running nuisance extraction
-
-        comm = (
-            "%s \"try general_extract_nuisance('%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s', %s, %s); catch ME, general_report_crash(ME); exit(1), end; exit\""
-            % (
-                mcommand,  # --- matlab command to run
-                f["bold_vol"],  # --- bold volume file to process
-                segfile,  # --- aseg or aparc file
-                f["bold1_brain_mask"],  # --- bold brain mask
-                d["s_bold_mov"],  # --- functional/movement subfolder
-                d["s_nuisance"],  # --- roi/nuisance subfolder
-                options["wbmask"],  # --- mask to exclude ROI from WB
-                options["sessionroi"],  # --- a mask used to specify session specific WB
-                options["nroi"],  # --- additional nuisance regressors ROI
-                options["shrinknsroi"],  # --- shrink nuisance signal ROI
-                "true",
-            )
-        )  # --- verbosity
-
-        if options["print_command"] == "yes":
-            log.pipeline_command(comm, title="Running:")
-
-        runit = True
-        if os.path.exists(f["bold_nuisance"]):
-            report["bolddone"] += 1
-            runit = False
-        endlog, status, failed = _run_external(
-            log,
-            options,
-            f["bold_nuisance"],
-            comm,
-            "... running matlab general_extract_nuisance on %s" % (f["bold_vol"]),
-            overwrite=overwrite,
-            thread=sinfo["id"],
-            remove=options["log"] == "remove",
-            task=options["command_ran"],
-            logfolder=options["comlogs"],
-            logtags=[
-                options["bold_variant"],
-                options["logtag"],
-                "B%d" % boldinfo["bold_number"],
-            ],
-            shell=True,
-        )
-        status = pc.check_for_file(
-            f["bold_nuisance"],
-            bad=f"Matlab/Octave has failed preprocessing BOLD using command: {comm}",
-            bad_level="error",
-            _log=log,
-        )
-
-        if runit and status:
-            report["boldok"] += 1
-        elif runit:
-            report["boldfail"] += 1
-
-    except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-        log.raw(str(errormessage))
-        report["boldfail"] += 1
-    except Exception:
-        log.error(f"Unknown error occured: \n...................................\n{traceback.format_exc()}...................................\n")
-        report["boldfail"] += 1
-
-    return {"r": log.text, "report": report}
-
-
-# -> @register_command(
-#        description="Preprocesses single BOLD images.",
-#         type="processing.session")
-def preprocess_bold(sinfo, options, overwrite=False, thread=0):
-    """
-    ``preprocess_bold [... processing options]``
-
-    Prepares BOLD files for further analysis. It performs spatial smoothing,
-    temporal filtering, removal of nuisance signals and complex modeling of
-    events. It is to be used when processing individual bold files. When
-    processing multiple bold files from a session for task-related analysis,
-    please use the ``preprocess_conc`` command instead.
-
-    Description:
-        Prepares BOLD files for further analysis. It performs spatial smoothing,
-        temporal filtering, removal of nuisance signals and complex modeling of
-        events. It is to be used when processing individual bold files. When
-        processing multiple bold files from a session for task-related analysis,
-        please use the ``preprocess_conc`` command instead.
-
-        preprocess_bold is a complex general purpose command implementing spatial
-        and temporal filtering, and multiple regression (GLM) to enable both
-        preprocessing and denoising of BOLD files for further analysis, as well as
-        complex activation modeling that creates GLM files for second-level
-        analyses. The function enables the following actions:
-
-        - spatial smoothing (3D or 2D for cifti files)
-        - temporal filtering (high-pass, low-pass)
-        - removal of nuisance signal
-        - complex modeling of events.
-
-        The function makes use of a number of files and accepts a long list of
-        arguments that make it very powerful and flexible but also require care in
-        its use. What follows is a detailed documentation of its actions and
-        parameters organized by actions in the order they would be most commonly
-        done. Use and parameter description will be intertwined.
-
-    ..  qx_command:
-        type: processing.session
-
-    Parameters:
-        --batchfile (str, default ''):
-            The batch.txt file with all the session information.
-
-        --sessionsfolder (str, default '.'):
-            The path to the study/sessions folder, where the imaging  data is
-            supposed to go.
-
-        --parsessions (str, default 1):
-            How many sessions to run in parallel.
-
-        --overwrite (str, default 'no'):
-            Whether to overwrite existing data (yes) or not (no). Note that
-            previous data is deleted before the run, so in the case of a failed
-            command run, previous results are lost.
-
-        --boldname (str, default 'bold'):
-            The default name of the bold files in the images folder.
-
-        --image_target (str, default 'nifti'):
-            The target format to work with, one of '4dfp', 'nifti', 'dtseries'
-            or 'ptseries'.
-
-        --logfolder (str, default ''):
-            The path to the folder where logs are to be stored,
-            if other than default.
-
-        --log (str, default 'keep'):
-            Whether to keep ('keep') or remove ('remove') the temporary logs
-            once jobs are completed. A comlog that recorded an error is kept
-            whichever is asked for, and a removed one still leaves its
-            completion status in the run log.
-
-        --comlog_folders (str, default ''):
-            A comma or pipe ('|') separated list of locations the comlogs are
-            placed in. The comlog is created at the first provided location and
-            then linked or copied to the others. The valid locations are:
-
-            - 'study'   (for the default:
-              ``<study>/processing/logs/comlogs`` location)
-            - 'session' (for ``<sessionid>/logs/comlogs``)
-            - 'hcp'     (for ``<hcp_folder>/logs/comlogs``)
-            - <path>  (for an arbitrary directory).
-
-            When empty, the value from the settings file is used, which
-            defaults to 'study'.
-
-        --bolds (str, default 'all'):
-            A pipe ('|') separated list of bold names to process.
-
-        --event_file (str, default ''):
-            A pipe ('|') separated list of fidl names to use, that
-            matches the bold list.
-
-        --bold_actions (str, default 's,h,r,c,l'):
-            A string specifying which actions, and in what sequence
-            to perform.
-
-        --nifti_tail (str, default ''):
-            The tail of NIfTI volume images to use.
-
-        --cifti_tail (str, default ''):
-            The tail of CIFTI images to use.
-
-        --bold_prefix (str, default ''):
-            An optional prefix to place in front of processing name extensions
-            in the resulting files, e.g. bold3<bold_prefix>_s_hpss.nii.gz.
-
-        --bold_variant (str, default detailed below):
-            Optional variant of HCP BOLD preprocessing. If specified, the BOLD
-            images in `images/functional<bold_variant>` will be processed.
-
-        --img_suffix (str, default ''):
-            Specifies a suffix for 'images' folder to enable support for
-            multiple parallel workflows. Empty if not used.
-
-    Output files:
-        This step results in the following files (if requested):
-
-        - residual image::
-
-            <root>_res-<regressors><glm name>.<ext>
-
-        - GLM image::
-
-            <bold name><bold tail>_<event root>_res-<regressors><glm name>_Bcoeff.<ext>
-
-        - text GLM regressor matrix::
-
-            glm/<bold name><bold tail>_GLM-X_<event root>_res-<regressors><glm name>.txt
-
-        - image of a regressor matrix::
-
-            ``glm/<bold name><bold tail>_GLM-X_<event root>_res-<regressors><glm name>.png``
-
-
-    Notes:
-        List of bold files specify, which types of bold files are to be
-        processed, as they are specified in the batch.txt file. An example of a
-        list of bolds in batch.txt would be::
-
-            07: bold1:blink       :BOLD blink 3mm 48 2.5s
-            08: bold2:flanker     :BOLD flanker 3mm 48 2.5s
-            09: bold3:EC          :BOLD EC 3mm 48 2.5s
-            10: bold4:mirror      :BOLD mirror 3mm 48 2.5s
-            11: bold5:rest        :RSBOLD 3mm 48 2.5s
-
-        With --bolds set to "blink|EC|rest", bold1, 3, and 5 would be processed.
-        If it were set to "all", all would be processed. As each bold gets
-        processed independently and only one fidl file can be specified, you
-        are advised to use preprocess_conc when regressing task structure, and
-        only use preprocess_bold for resting state data. If you would still use
-        like to regress out events specified in a fidl file. They would neet to
-        be named as [<session id>_]<boldname>_<image_target>_<fidl name>.fidl.
-        In the case of cifti files, image_target is composed of
-        <cifti_tail>_cifti. If the files are not present in the relevant
-        individual sessions's folders, they are searched for in the
-        <sessionsfolder>/inbox/events folder. In that case the "<session id>_"
-        is not optional but required.
-
-        The --bold_actions parameter specifies the actions, denoted by a single
-        letter, that will be executed in the sequence listed:
-
-        --m
-            Motion scrubbing.
-        --s
-            Spatial smoothing.
-        --h
-            High-pass filtering.
-        --r
-            Regression (nuisance and/or task) with an optional number 0, 1, or 2
-            specifying the type of regression to use (see REGRESSION below).
-        --l
-            Low-pass filtering.
-        --c
-            Save beta coefficients as well.
-
-        So the default 's,h,r,c,l' bold_actions parameter would lead to the
-        files first being smoothed, then high-pass filtered. Next a regression
-        step would follow in which nuisance signal and/or task related signal
-        would be estimated and regressed out, then the related beta estimates
-        would be saved. Lastly the BOLDs would be also low-pass filtered.
-
-        Use:
-            preprocess_bold is a complex command initially used to prepare BOLD
-            files for further functional connectivity analysis. The function
-            enables the following actions:
-
-            - spatial smoothing (3D or 2D for cifti files)
-            - temporal filtering (high-pass, low-pass)
-            - removal of nuisance signal and task structure
-
-            The function makes use of a number of files and accepts a long list
-            of arguments that make it very powerful and flexible but also
-            require care in its use. What follows is a detailed documentation
-            of its actions and parameters organized by actions in the order
-            they would be most commonly done. Use and parameter description
-            will be intertwined.
-
-        Scrubbing:
-            The command either makes use of scrubbing information or performs
-            scrubbing comuputation on its own (when 'm' is part of the
-            command). In the latter case, all the scrubbing parameters need to
-            be specified:
-
-            --mov_radius (int, default 50):
-                Estimated head radius (in mm) for computing frame displacement
-                statistics.
-            --mov_fd (float, default 0.5):
-                Frame displacement threshold (in mm) to use for identifying bad
-                frames.
-            --mov_dvars (float, default 3.0):
-                The (mean normalized) dvars threshold to use for identifying bad
-                frames.
-            --mov_dvarsme (float, default 1.5):
-                The (median normalized) dvarsm threshold to use for identifying
-                bad frames.
-            --mov_after (int, default 0):
-                How many frames after each frame identified as bad to also
-                exclude from further processing and analysis.
-            --mov_before (int, default 0):
-                How many frames before each frame identified as bad to also
-                exclude from further processing and analysis.
-            --mov_bad (str, default 'udvarsme'):
-                Which criteria to use for identification of bad frames.
-
-            Criteria for identification of bad frames can be one out of:
-
-            --mov
-                Frame displacement threshold (fdt) is exceeded.
-            --dvars
-                Image intensity normalized root mean squared error (RMSE)
-                threshold (dvarsmt) is exceeded.
-            --dvarsme
-                Median normalised RMSE (dvarsmet) threshold is exceeded.
-            --idvars
-                Both fdt and dvarsmt are exceeded (i for intersection).
-            --uvars
-                Either fdt or dvarsmt are exceeded (u for union).
-            --idvarsme
-                Both fdt and dvarsmet are exceeded.
-            --udvarsme
-                Either fdt or udvarsmet are exceeded.
-
-            For more detailed description please see wiki entry on Movement
-            scrubbing.
-
-            In any case, if scrubbing was done beforehand or as a part of this
-            command, one has to specify, how the scrubbing information is used:
-
-            --pignore
-                String describing how to deal with bad frames.
-
-            The string has the following format::
-
-                'hipass:<filtering opt.>|regress:<regression opt.>|lopass:<filtering opt.>'
-
-            Filtering options are:
-
-            --keep
-                Keep all the bad frames unchanged.
-            --linear
-                Replace bad frames with linear interpolated values based on
-                neighboring good frames.
-            --spline
-                Replace bad frames with spline interpolated values based on
-                neighboring good frames.
-
-            To prevent artifacts present in bad frames to be temporarily spread,
-            use either 'linear' or 'spline' options.
-
-            Regression options are:
-
-            --keep
-                Keep the bad frames and use them in the regression.
-            --ignore
-                Exclude bad frames from regression.
-            --mark
-                Exclude bad frames from regression and mark the bad frames as
-                NaN.
-            --linear
-                Replace bad frames with linear interpolated values based on
-                neighboring good frames.
-            --spline
-                Replace bad frames with spline interpolated values based on
-                neighboring good frames.
-
-            Please note that when the bad frames are not kept, the original
-            values will be retained in the residual signal. In this case they
-            have to be excluded or ignored also in all following analyses,
-            otherwise they can be a significant source of artifacts.
-
-        Spatial smoothing:
-            Volume smoothing:
-                For volume formats the images will be smoothed using the
-                img_smooth_3d nimage method. For cifti format the smooting will
-                be done by calling the relevant wb_command command. The
-                smoothing parameters are:
-
-                --voxel_smooth (int, default 1):
-                    Gaussian smoothing FWHM in voxels.
-                --smooth_mask (str, default false):
-                    Whether to smooth only within a mask, and what mask to use
-                    (nonzero|brainsignal|brainmask|<filename>|false).
-                --dilate_mask (str, default false):
-                    Whether to dilate the image after masked smoothing and what
-                    mask to use
-                    (nonzero|brainsignal|brainmask|same|<filename>|false).
-
-                If a smoothing mask is set, only the signal within the specified
-                mask will be used in the smoothing. If a dilation mask is set,
-                after smoothing within a mask, the resulting signal will be
-                constrained / dilated to the specified dilation mask.
-
-                For both parameters the possible options are:
-
-                --nonzero
-                    Mask will consist of all the nonzero voxels of the first
-                    BOLD frame.
-                --brainsignal
-                    Mask will consist of all the voxels that are of value 300 or
-                    higher in the first BOLD frame (this gave a good coarse
-                    brain mask for images intensity normalized to mode 1000 in
-                    the NIL preprocessing stream).
-                --brainmask
-                    Mask will be the actual bet extracted brain mask based on
-                    the first BOLD frame (generated using in the
-                    creatBOLDBrainMasks command).
-                --filename
-                    All the non-zero voxels in a specified volume file will be
-                    used as a mask.
-                --false
-                    No mask will be used.
-                --same
-                    Only for dilate_mask, the mask used will be the same as
-                    smoothing mask.
-
-            Cifti smoothing:
-                For cifti format images, smoothing will be run using wb_command.
-                The following parameters can be set:
-
-                --surface_smooth (float, default 2.0):
-                    FWHM for Gaussian surface smoothing in mm.
-                --volume_smooth (float, default 2.0):
-                    FWHM for Gaussian volume smoothing in mm.
-                --framework_path (str, default ''):
-                    The path to framework libraries on the Mac system. No need
-                    to use it currently if installed correctly.
-                --wb_command_path (str, default ''):
-                    The path to the wb_command executive. No need to use it
-                    currently if installed correctly.
-
-                Results:
-                    The resulting smoothed files are saved with '_s' added to
-                    the BOLD root filename.
-
-        Temporal filtering:
-            Temporal filtering is accomplished using img_filter nimage method.
-            The code is adopted from the FSL C++ code enabling appropriate
-            handling of bad frames (as described above - see SCRUBBING). The
-            parameters are:
-
-            --hipass_filter (float, default 0.008):
-                The frequency for high-pass filtering in Hz.
-            --lopass_filter (float, default 0.09):
-                The frequency for low-pass filtering in Hz.
-
-            Please note that the values finally passed to img_filter method are
-            the respective sigma values computed from the specified frequencies
-            and TR.
-
-            Filtering of nuisance signal, movement, task, and events
-            Besides data, nuisance signal, motion parameters, and event
-            regressors can be filtered as well. What to filter beside data can
-            be specified by a comma separated list using the following
-            parameters:
-
-            --hipass_do (str, default 'nuisance')
-                What to high-pass filter besides data – options are: nuisance,
-                movement, events, task. Default is 'nuisance'.
-
-            --lopass_do (str, default 'nuisance,movement,task,events')
-                What to lo-pass filter besides data – options are: nuisance,
-                movement, events, task. Default is 'nuisance, movement, task,
-                events'.
-
-            Note that 'events' refers to regressors created based on events
-            as specified in the fidl file, whereas 'task' refers to a task
-            matrix that is passed directy in the matlab function call.
-
-            Results:
-                The resulting filtered files are saved with '_hpss' or '_bpss'
-                added to the BOLD root filename for high-pass and low-pass
-                filtering, respectively.
-
-        Regression:
-            Regression is a complex step in which GLM is used to estimate the
-            beta weights for the specified nuisance regressors and events. The
-            resulting beta weights are then stored in a GLM file (a regular
-            file with additional information on the design used) and residuals
-            are stored in a separate file. This step can therefore be used for
-            two purposes: (1) to remove nuisance signal and event structure
-            from BOLD files, removing unwanted potential sources of correlation
-            for further functional connectivity analyses, and (2) to get task
-            beta estimates for further activation analyses. The following
-            parameters are used in this step:
-
-            --bold_nuisance (str, default 'm,m1d,mSq,m1dSq,V,WM,WB,1d'):
-                A comma separated list of regressors to include in GLM.
-                Possible values are:
-
-                - 'm'     ... motion parameters
-                - 'm1d'   ... first derivative of motion parameters
-                - 'mSq'   ... squared motion parameters
-                - 'm1dSq' ... squared first derivative of motion parameters
-                - 'V'     ... ventricles signal
-                - 'WM'    ... white matter signal
-                - 'WB'    ... whole brain signal
-                - '1d'    ... first derivative of above nuisance signals
-                - 'e'     ... events listed in the provided fidl files.
-
-            --event_string (str, default ''):
-                A string describing, how to model the events listed in the
-                provided fidl files.
-            --glm_matrix (str, default 'none'):
-                Whether to save the GLM matrix as a text file ('text'), a png
-                image file ('image'), both ('both') or not ('none').
-            --glm_results (str, default 'c,r')
-                A string  which of the GLM analysis results are saved.
-                Possible values are:
-
-                - 'c'   ... Saving of resulting beta coefficients.
-                - 'z'   ... Saving of resulting z-scores of beta coefficients.
-                - 'p'   ... Saving of session-level coefficient p-values.
-                - 'se'  ... Saving of standard errors of beta coefficients.
-                - 'r'   ... Saving of resulting residuals of the GLM.
-                - 'all' ... Saving all of the results above.
-
-            --glm_name (str, default ''):
-                An additional name to add to the residuals and GLM files to
-                distinguish between different possible models used.
-
-        GLM modeling:
-            There are two important variables that affect the exact GLM
-            model used to estimate nuisance and task beta coefficients and
-            regress them from the signal. The first is the optional number
-            following the 'r' command in the --bold_actions parameter.
-            There are three options:
-
-            - '0' ... Estimate nuisance regressors for each bold file
-                separately, however, model events across all bold files (the
-                default if no number is) specified.
-            - '1' ... Estimate both nuisance regressors and task regressors
-                for each bold run separately.
-            - '2' ... Estimate both nuisance regressors as well as task
-                regressors across all bold runs.
-
-            The second key variable is the event string provided by the
-            event_string parameter. The event string is a pipe ('|')
-            separated list of regressor specifications. The possibilities
-            are discussed below:
-
-            Unassumed Modeling:
-                ::
-
-                    <fidl code>:<length in frames>
-
-                Where <fidl code> is the code for the event used in the fidl
-                file, and <length in frames> specifies, for how many frames
-                of the bold run (since the onset of the event) the event
-                should be modeled.
-
-            Assumed Modeling:
-                ::
-
-                    <fidl code>:<hrf>[-run|-uni][:<length>]
-
-                Where <fidl code> is the same as above, <hrf> is the type of
-                the hemodynamic response function to use, '-run' and '-uni'
-                specify how the regressor should be normalized, and
-                <length> is an optional parameter, with its value dependent
-                on the model used. The allowed <hrf> are:
-
-                - 'boynton' ... uses the Boynton HRF
-                - 'SPM'     ... uses the SPM double gaussian HRF
-                - 'u'       ... unassumed (see above)
-                - 'block'   ... block response.
-
-                For the first two, the <length> parameter is optional and
-                would override the event duration information provided in
-                the fidl file. For 'u' the length is the same as in
-                previous section: the number of frames to model. For
-                'block' length should be two numbers separated by a colon
-                (e.g. 2:9) that specify the start and end offset (from the
-                event onset) to model as a block.
-
-                Assumed HRF regressors normalization:
-                    hrf_types `boynton` and `SPM` can be marked with an
-                    additional flag denoting how to normalize the
-                    regressor.
-
-                    In case of `<hrf function>-uni`, e.g. 'boynton-uni' or
-                    'SPM-uni', the HRF function will be normalized to have
-                    the area under the curve equal to 1. This ensures
-                    uniform and universal, scaling of the resulting
-                    regressor across all event lengths. In addition, the
-                    scaling is not impacted by weights (e.g. behavioral
-                    coregressors), which in turn ensures that the weights
-                    are not scaled.
-
-                    In case of `<hrf function>-run`, e.g. `boynton-run` or
-                    `SPM-run`, the resulting regressor is normalized to
-                    amplitude of 1 within each bold run separately. This
-                    can result in different scaling of regressors with
-                    different durations, and of the same regressor across
-                    different runs. Scaling in this case is performed after
-                    the signal is weighted, so in effect the scaling of
-                    weights (e.g. behavioral regressors), can differ across
-                    bold runs.
-
-                    The flag can be abbreviated to '-r' and '-u'. If not
-                    specified, '-run' will be assumed (the default might
-                    change).
-
-        Naming And Behavioral Regressors:
-            Each of the above (unassumed and assumed modeling
-            specification) can be followed by a ">" (greater-than
-            character), which signifies additional information in the
-            form::
-
-                <name>[:<column>[:<normalization_span>[:<normalization_method>]]]
-
-            --name
-                The name of the resulting regressor.
-            --column
-                The number of the additional behavioral regressors
-                column in the fidl file (1-based) to use as a weight
-                for the regressors.
-            --normalization_span
-                Whether to normalize the behavioral weight within a
-                specific event type ('within') or across all events
-                ('across'). [within]
-            --normalization_method
-                The method to use for normalization. Options are:
-
-                - 'z'    (compute Z-score)
-                - '01'   (normalize to fixed range 0 to 1)
-                - '-11'  (normalize to fixed range -1 to 1)
-                - 'none' (use weights as provided in fidl file)
-
-                Example string::
-
-                    'block:boynton|target:9|target:9>target_rt:1:within:z'
-
-                This would result in three sets of task regressors: one
-                assumed task regressor for the sustained activity across
-                the block, one unassumed task regressor set spanning 9
-                frames that would model the presentation of the target, and
-                one behaviorally weighted unassumed regressor that would
-                for each frame estimate the variability in response as
-                explained by the reaction time to the target.
-
-        Results:
-            This step results in the following files (if requested):
-
-            - residual image (``<root>_res-<regressors>.<ext>``)
-            - GLM coefficient image (``<root>_res-<regressors>_coeff.<ext>``)
-
-            If you want more specific GLM results and information, please use
-            preprocess_conc command.
-
-    Examples:
-        ::
-
-            qunex preprocess_bold \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --parsessions=10 \\
-                --bolds=rest \\
-                --bold_actions="s,h,r,c,l" \\
-                --bold_nuisance="m,V,WM,WB,1d" \\
-                --mov_bad=udvarsme \\
-                --pignore="hipass=linear|regress=ignore|lopass=linear"
-    """
-    log = ReportLog()
-
-    pc.do_options_check(options, sinfo, "preprocess_bold")
-
-    log.rule()
-    log.info(f"Session id: {sinfo['id']} \n[started on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}]")
-    log.info(f"Preprocessing {', '.join(options['bolds'].split('|'))} BOLD files as specified in --bolds.")
-    log.info(f"Files in 'images{options['img_suffix']}/functional{options['bold_variant']} will be processed.")
-    log.action("Running", "Preprocessing bold runs ...", options["run"], level="info")
-
-    report = {
-        "done": [],
-        "processed": [],
-        "failed": [],
-        "ready": [],
-        "not ready": [],
-        "skipped": [],
-    }
-
-    bolds, bskip, report["boldskipped"] = pc.use_or_skip_bold(sinfo, options, _log=log)
-    report["skipped"] = [str(binfo["bold_number"]) for binfo in bskip]
-
-    parelements = options["parelements"]
-    log.info(f"Processing {parelements} BOLDs in parallel")
-
-    if parelements == 1:  # serial execution
-        for b in bolds:
-            # process
-            result = execute_preprocess_bold(sinfo, options, overwrite, b)
-
-            # merge r
-            log.raw(result["r"])
-
-            # merge report
-            temp_report = result["report"]
-            report["done"] += temp_report["done"]
-            report["processed"] += temp_report["processed"]
-            report["failed"] += temp_report["failed"]
-            report["ready"] += temp_report["ready"]
-            report["not ready"] += temp_report["not ready"]
-    else:  # parallel execution
-        # create a multiprocessing Pool
-        process_pool_executor = ProcessPoolExecutor(parelements)
-        # process
-        f = partial(execute_preprocess_bold, sinfo, options, overwrite)
-        results = process_pool_executor.map(f, bolds)
-
-        # merge r and report
-        for result in results:
-            log.raw(result["r"])
-            temp_report = result["report"]
-            report["done"] += temp_report["done"]
-            report["processed"] += temp_report["processed"]
-            report["failed"] += temp_report["failed"]
-            report["ready"] += temp_report["ready"]
-            report["not ready"] += temp_report["not ready"]
-
-    log.blank()
-    log.info(f"Bold preprocessing completed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-    log.rule()
-    if options["run"] == "run":
-        rstatus = (
-            "bolds: %d ready [%s], %d not ready [%s], %d already processed [%s], %d ran ok [%s], %d failed [%s], %d skipped [%s]"
-            % (
-                len(report["ready"]),
-                " ".join(report["ready"]),
-                len(report["not ready"]),
-                " ".join(report["not ready"]),
-                len(report["done"]),
-                " ".join(report["done"]),
-                len(report["processed"]),
-                " ".join(report["processed"]),
-                len(report["failed"]),
-                " ".join(report["failed"]),
-                len(report["skipped"]),
-                " ".join(report["skipped"]),
-            )
-        )
-    else:
-        rstatus = (
-            "bolds: %d ready [%s], %d not ready [%s], %d already processed [%s], %d skipped [%s]"
-            % (
-                len(report["ready"]),
-                " ".join(report["ready"]),
-                len(report["not ready"]),
-                " ".join(report["not ready"]),
-                len(report["done"]),
-                " ".join(report["done"]),
-                len(report["skipped"]),
-                " ".join(report["skipped"]),
-            )
-        )
-
-    # print r
-    return log.result(
-        rstatus, len(report["not ready"]) + len(report["failed"]), sinfo["id"]
-    )
-
-
-def execute_preprocess_bold(sinfo, options, overwrite, boldinfo):
-
-    # prepare return variables
-    log = ReportLog()
-    report = {"done": [], "processed": [], "failed": [], "ready": [], "not ready": []}
-
-    boldnum = str(boldinfo["bold_number"])
-
-    log.raw("\n\nWorking on: " + boldinfo["name"] + " ...")
-
-    try:
-        # --- define the tail
-
-        options["bold_tail"] = options["nifti_tail"]
-        if options["image_target"] in ["cifti", "dtseries", "ptseries"]:
-            options["bold_tail"] = options["cifti_tail"]
-
-        # --- filenames and folders
-
-        f = pc.get_file_names(sinfo, options)
-        f.update(pc.get_bold_file_names(sinfo, boldinfo["name"], options))
-        d = pc.get_session_folders(sinfo, options)
-
-        # --- check for data availability
-
-        log.detail("checking for data")
-        status = True
-
-        # --- movement
-        if "r" in options["bold_actions"] and (
-            "m" in options["bold_nuisance"] or "m" in options["bold_actions"]
-        ):
-            status = pc.check_for_file(f["bold_mov"],
-                "movement data present",
-                f"movement data missing [{f['bold_mov']}]",
-                status=status,
-                _log=log,
-            )
-
-        # --- bold stats
-        if "m" in options["bold_actions"]:
-            status = pc.check_for_file(f["bold_stats"],
-                "bold statistics data present",
-                f"bold statistics data missing [{f['bold_stats']}]",
-                status=status,
-                _log=log,
-            )
-
-        # --- bold scrub
-        if any([e in options["pignore"] for e in ["linear", "spline", "ignore"]]):
-            status = pc.check_for_file(f["bold_scrub"],
-                "bold scrubbing data present",
-                f"bold scrubbing data missing [{f['bold_scrub']}]",
-                status=status,
-                _log=log,
-            )
-
-        # --- check for files if doing regression
-
-        if "r" in options["bold_actions"]:
-            # --- nuisance data
-            if any([e in options["bold_nuisance"] for e in ["V", "WM", "WB"]]):
-                status = pc.check_for_file(f["bold_nuisance"],
-                    "bold nuisance signal data present",
-                    f"bold nuisance signal data missing [{f['bold_nuisance']}]",
-                    status=status,
-                    _log=log,
-                )
-
-            # --- event
-            if "e" in options["bold_nuisance"]:
-                status = pc.check_for_file(f["bold_event"],
-                    "event data present",
-                    f"even data missing [{f['bold_event']}]",
-                    status=status,
-                    _log=log,
-                )
-
-        # --- bold
-        status = pc.check_for_file(f["bold"],
-            "bold data present",
-            f"bold data missing [{f['bold']}]",
-            status=status,
-            _log=log,
-        )
-
-        # --- results
-        already_done = pc.check_for_file(f["bold_final"], "result present", _log=log)
-
-        # --- check
-        if not status:
-            log.error("Files missing, skipping this bold run!")
-            report["not ready"].append(boldnum)
-            return {"r": log.text, "report": report}
-        else:
-            report["ready"].append(boldnum)
-
-        if already_done:
-            report["done"].append(boldnum)
-
-        # --- run matlab preprocessing script
-
-        if overwrite:
-            boldow = "true"
-        else:
-            boldow = "false"
-
-        scrub = (
-            "radius:%(mov_radius)d|fdt:%(mov_fd).2f|dvarsmt:%(mov_dvars).2f|dvarsmet:%(mov_dvarsme).2f|after:%(mov_after)d|before:%(mov_before)d|reject:%(mov_bad)s"
-            % (options)
-        )
-        opts = (
-            "boldname=%(boldname)s|surface_smooth=%(surface_smooth)f|volume_smooth=%(volume_smooth)f|voxel_smooth=%(voxel_smooth)f|hipass_filter=%(hipass_filter)f|lopass_filter=%(lopass_filter)f|framework_path=%(framework_path)s|wb_command_path=%(wb_command_path)s|smooth_mask=%(smooth_mask)s|dilate_mask=%(dilate_mask)s|glm_matrix=%(glm_matrix)s|glm_residuals=%(glm_residuals)s|glm_results=%(glm_results)s|glm_name=%(glm_name)s|bold_tail=%(bold_tail)s|ref_bold_tail=%(nifti_tail)s|bold_variant=%(bold_variant)s|img_suffix=%(img_suffix)s"
-            % (options)
-        )
-
-        mcomm = (
-            "fc_preprocess('%s', %s, %d, '%s', '%s', %s, '%s', %f, '%s', '%s', %s, '%s', '%s', '%s', '%s')"
-            % (
-                d["s_base"],  # --- sessions folder
-                boldnum,  # --- number of bold file to process
-                options[
-                    "omit"
-                ],  # --- number of frames to skip at the start of each run
-                options[
-                    "bold_actions"
-                ],  # --- which steps to perform (s, h, r, c, p, p, z)
-                options[
-                    "bold_nuisance"
-                ],  # --- what to regress (m, m1d, mSq, m1dSq, V, WM, WB, d, t, e, 1d)
-                "[]",  # --- matrix of task regressors
-                options["event_file"],  # --- fidl file to be used
-                options["tr"],  # --- TR of the data
-                options[
-                    "event_string"
-                ],  # --- event string specifying what and how of the task to regress
-                options["bold_prefix"],  # --- prefix to the bold files
-                boldow,  # --- whether to overwrite the existing files
-                gi.get_img_format(
-                    f["bold_final"]
-                ),  # --- what file extension to expect and use (e.g. '.nii', .'.4dfp.img')
-                scrub,  # --- scrub parameters
-                options[
-                    "pignore"
-                ],  # --- how to deal with bad frames ('hipass:keep/linear/spline|regress:keep/ignore|lopass:keep/linear/spline')
-                opts,
-            )
-        )  # --- additional options
-
-        comm = '%s "try %s; catch ME, general_report_crash(ME); exit(1), end; exit"' % (
-            mcommand,
-            mcomm,
-        )
-
-        # r += '\n ... running: %s' % (comm)
-        if already_done and not overwrite:
-            log.raw("\n\nProcessing already completed! Set overwrite to yes to redo processing!\n")
-        else:
-            if options["print_command"] == "yes":
-                log.raw("\n\nRunning\n" + comm + "\n")
-            endlog, status, failed = _run_external(
-                log,
-                options,
-                f["bold_final"],
-                comm,
-                "running matlab/octave fc_preprocess on %s bold %s"
-                % (d["s_bold"], boldnum),
-                overwrite=overwrite,
-                thread=sinfo["id"],
-                remove=options["log"] == "remove",
-                task=options["command_ran"],
-                logfolder=options["comlogs"],
-                logtags=[
-                    options["bold_variant"],
-                    options["glm_name"],
-                    options["logtag"],
-                    "B%s" % (boldnum),
-                ],
-                shell=True,
-            )
-            # the check reads what the guarded call would have written, so it
-            # belongs inside the branch that does the work
-            if options["run"] == "run":
-                status = pc.check_for_file(
-                    f["bold_final"],
-                    bad=f"Matlab/Octave has failed preprocessing BOLD using command: \n---> {mcomm}\n",
-                    bad_level="error",
-                    _log=log,
-                )
-                if status:
-                    report["processed"].append(boldnum)
-                else:
-                    report["failed"].append(boldnum)
-    except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-        log.raw(str(errormessage))
-        report["failed"].append(boldnum)
-    except Exception:
-        log.error(f"Unknown error occured: \n...................................\n{traceback.format_exc()}...................................\n")
-        time.sleep(5)
-        report["failed"].append(boldnum)
-
-    return {"r": log.text, "report": report}
-
-
-def preprocess_conc(sinfo, options, overwrite=False, thread=0):
-    r"""
-    ``preprocess_conc [... processing options]``
-
-    Preproces conc bundles of BOLD images.
-
-    Description:
-        Performs spatial smoothing, temporal filtering, removal of nuisance signals
-        and complex modeling of events.
-
-        preprocess_conc is a complex general purpose command implementing spatial
-        and temporal filtering, and multiple regression (GLM) to enable both
-        preprocessing and denoising of BOLD files for further analysis, as well as
-        complex activation modeling that creates GLM files for second-level
-        analyses. The function enables the following actions:
-
-        - spatial smoothing (3D or 2D for cifti files)
-        - temporal filtering (high-pass, low-pass)
-        - removal of nuisance signal
-        - complex modeling of events.
-
-        The function makes use of a number of files and accepts a long list of
-        arguments that make it very powerful and flexible but also require care in
-        its use. What follows is a detailed documentation of its actions and
-        parameters organized by actions in the order they would be most commonly
-        done. Use and parameter description will be intertwined.
-
-    ..  qx_command:
-        type: processing.session
-
-
-
-    Parameters:
-        --batchfile (str, default ''):
-            The batch.txt file with all the session information.
-
-        --sessionsfolder (str, default '.'):
-            The path to the study/sessions folder, where the imaging  data is
-            supposed to go.
-
-        --parsessions (str, default 1):
-            How many sessions to run in parallel.
-
-        --overwrite (str, default 'no'):
-            Whether to overwrite existing data (yes) or not (no). Note that
-            previous data is deleted before the run, so in the case of a failed
-            command run, previous results are lost.
-
-        --boldname (str, default 'bold'):
-            The default name of the bold files in the images folder.
-
-        --image_target (str, default 'nifti'):
-            The target format to work with, one of '4dfp', 'nifti', 'dtseries'
-            or 'ptseries'.
-
-        --logfolder (str, default ''):
-            The path to the folder where logs are to be stored,
-            if other than default.
-
-        --log (str, default 'keep'):
-            Whether to keep ('keep') or remove ('remove') the temporary logs
-            once jobs are completed. A comlog that recorded an error is kept
-            whichever is asked for, and a removed one still leaves its
-            completion status in the run log.
-
-        --comlog_folders (str, default ''):
-            A comma or pipe ('|') separated list of locations the comlogs are
-            placed in. The comlog is created at the first provided location and
-            then linked or copied to the others. The valid locations are:
-
-            - 'study'   (for the default:
-              ``<study>/processing/logs/comlogs`` location)
-            - 'session' (for ``<sessionid>/logs/comlogs``)
-            - 'hcp'     (for ``<hcp_folder>/logs/comlogs``)
-            - <path>  (for an arbitrary directory).
-
-            When empty, the value from the settings file is used, which
-            defaults to 'study'.
-
-        --bolds (str, default 'all'):
-            A pipe ('|') separated list of conc names to process.
-
-        --event_file (str, default ''):
-            A pipe ('|') separated list of fidl names to use, that
-            matches the conc list.
-
-        --bold_actions (str, default 's,h,r,c,l'):
-            A string specifying which actions, and in what sequence
-            to perform.
-
-        --nifti_tail (str, default ''):
-            The tail of NIfTI volume images to use.
-
-        --cifti_tail (str, default ''):
-            The tail of CIFTI images to use.
-
-        --bold_prefix (str, default ''):
-            An optional prefix to place in front of processing name extensions
-            in the resulting files, e.g. bold3<bold_prefix>_s_hpss.nii.gz.
-
-        --bold_variant (str, default detailed below):
-            Optional variant of HCP BOLD preprocessing. If specified, the BOLD
-            images in `images/functional<bold_variant>` will be processed.
-
-        --img_suffix (str, default ''):
-            Specifies a suffix for 'images' folder to enable support for
-            multiple parallel workflows. Empty if not used.
-
-    Output files:
-        This step results in the following files (if requested):
-
-        - residual image::
-
-            <root>_res-<regressors><glm name>.<ext>
-
-        - GLM image::
-
-            <bold name><bold tail>_conc_<event root>_res-<regressors><glm name>_Bcoeff.<ext>
-
-        - text GLM regressor matrix::
-
-            glm/<bold name><bold tail>_GLM-X_<event root>_res-<regressors><glm name>.txt
-
-        - image of a regressor matrix::
-
-            ``glm/<bold name><bold tail>_GLM-X_<event root>_res-<regressors><glm name>.png``
-
-    Notes:
-        The --bolds and --event_file parameters provide names based on which the
-        the appropriate .conc and .fidl files are searched for. Both are first
-        searched for in `images<img_suffix>/functional<bold_variant>/concs` and
-        images<img_suffix>/functional<bold_variant>/events folders
-        respectively. There they would be named as `[<session
-        id>_]<boldname>_<image_target>_<conc name>.conc` and `[<session
-        id>_]<boldname>_<image_target>_<fidl name>.fidl`. If the files are not
-        present in the relevant individual session's folders, they are searched
-        for in the `<sessionsfolder>/inbox/events` and
-        `<sessionsfolder>/inbox/concs` folder. In that case the "<session id>_"
-        in the \*.fidl and \*.conc file name is not optional but required.
-
-        The --bold_actions parameter specifies the actions, denoted by a single
-        letter, that will be executed in the sequence listed:
-
-        --m     Motion scrubbing.
-        --s     Spatial smoothing.
-        --h     High-pass filtering.
-        --r     Regression (nuisance and/or task) with an optional number 0, 1,
-                or 2 specifying the type of regression to use (see REGRESSION
-                below).
-        --l     Low-pass filtering.
-
-        So the default 's,h,r,c,l' --bold_actions parameter would lead to the
-        files first being smoothed, then high-pass filtered. Next a regression
-        step would follow in which nuisance signal and/or task related signal
-        would be estimated and regressed out, then the related beta estimates
-        would be saved. Lastly the BOLDs would be also low-pass filtered.
-
-        Relative vs. absolute use of conc files:
-            If `conc_use` is set to relative (the default), then the only
-            information taken from the conc files will be the bold numbers. The
-            actual location of the bold files will be constructed from the
-            information on the location of the session's sesion folder present
-            in the batch file, and the `bold_variant` setting, whereas the
-            specific bold file name and file format (e.g. .nii.gz vs.
-            .dtseries.nii) to use will depend on `boldname`, `image_target`,
-            `nifti_tail` and `cifti_tail` settings. This allows for flexible
-            use of conc files. That is, the same conc files can be used for
-            NIfTI and CIFTI versions of bold files, across bold variants, and
-            even when the actual study location changes, e.g. when moving the
-            study from one server, volume or folder to another. In most cases
-            this use will be prefered.
-
-            If the information in the conc file is to be used literally, e.g. in
-            cases when you want to work with a specific preprocessed version of
-            the BOLD files, then `conc_use` should be set to `absolute`. In
-            this case both the specific location as well as the specific
-            filename specified in the conc file will be used exactly as
-            specified. In this case, do check and make sure that the
-            information in the conc file is valid and it matches with
-            `boldname` and `image_target` parameters, and that the `nifti_tail`
-            is specified correctly, as it will be used to obtain bold
-            statistics and nuisance information!
-
-        Scrubbing:
-            The command either makes use of scrubbing information or performs
-            scrubbing comuputation on its own (when 'm' is part of the
-            command). In the latter case, all the scrubbing parameters need to
-            be specified:
-
-            --mov_radius (int, default 50):
-                Estimated head radius (in mm) for computing frame displacement
-                statistics.
-            --mov_fd (float, default 0.5):
-                Frame displacement threshold (in mm) to use for identifying bad
-                frames.
-            --mov_dvars (float, default 3.0):
-                The (mean normalized) dvars threshold to use for identifying bad
-                frames.
-            --mov_dvarsme (float, default 1.5):
-                The (median normalized) dvarsm threshold to use for identifying
-                bad frames.
-            --mov_after (int, default 0):
-                How many frames after each frame identified as bad to also
-                exclude from further processing and analysis.
-            --mov_before (int, default 0):
-                How many frames before each frame identified as bad to also
-                exclude from further processing and analysis.
-            --mov_bad (str, default 'udvarsme'):
-                Which criteria to use for identification of bad frames.
-
-            Criteria for identification of bad frames can be one out of:
-
-            --mov
-                Frame displacement threshold (fdt) is exceeded.
-            --dvars
-                Image intensity normalized root mean squared error (RMSE)
-                threshold (dvarsmt) is exceeded.
-            --dvarsme
-                Median normalised RMSE (dvarsmet) threshold is exceeded.
-            --idvars
-                Both fdt and dvarsmt are exceeded (i for intersection).
-            --uvars
-                Either fdt or dvarsmt are exceeded (u for union).
-            --idvarsme
-                Both fdt and dvarsmet are exceeded.
-            --udvarsme
-                Either fdt or dvarsmet are exceeded.
-
-            For more detailed description please see wiki entry on Movement
-            scrubbing.
-
-            In any case, if scrubbing was done beforehand or as a part of this
-            command, one has to specify, how the scrubbing information is used:
-
-            --pignore
-                String describing how to deal with bad frames.
-
-            The string has the following format::
-
-                'hipass:<filtering opt.>|regress:<regression opt.>|lopass:<filtering opt.>'
-
-            Filtering options are:
-
-            --keep
-                Keep all the bad frames unchanged.
-            --linear
-                Replace bad frames with linear interpolated values based on
-                neighboring good frames.
-            --spline
-                Replace bad frames with spline interpolated values based on
-                neighboring good frames.
-
-            To prevent artifacts present in bad frames to be temporarily spread,
-            use either 'linear' or 'spline' options.
-
-            Regression options are:
-
-            --keep
-                Keep the bad frames and use them in the regression.
-            --ignore
-                Exclude bad frames from regression.
-            --mark
-                Exclude bad frames from regression and mark the bad frames as
-                NaN.
-            --linear
-                Replace bad frames with linear interpolated values based on
-                neighboring good frames.
-            --spline
-                Replace bad frames with spline interpolated values based on
-                neighboring good frames.
-
-            Please note that when the bad frames are not kept, the original
-            values will be retained in the residual signal. In this case they
-            have to be excluded or ignored also in all following analyses,
-            otherwise they can be a significant source of artifacts.
-
-        Spatial smoothing:
-            Volume smoothing:
-                For volume formats the images will be smoothed using the
-                img_smooth_3d nimage method. For cifti format the smooting will
-                be done by calling the relevant wb_command command. The
-                smoothing parameters are:
-
-                --voxel_smooth (int, default 1):
-                    Gaussian smoothing FWHM in voxels.
-                --smooth_mask (str, default false):
-                    Whether to smooth only within a mask, and what mask to use
-                    (nonzero|brainsignal|brainmask|<filename>|false).
-                --dilate_mask (str, default false):
-                    Whether to dilate the image after masked smoothing and what
-                    mask to use
-                    (nonzero|brainsignal|brainmask|same|<filename>|false).
-
-                If a smoothing mask is set, only the signal within the specified
-                mask will be used in the smoothing. If a dilation mask is set,
-                after smoothing within a mask, the resulting signal will be
-                constrained / dilated to the specified dilation mask.
-
-                For both parameters the possible options are:
-
-                --nonzero
-                    Mask will consist of all the nonzero voxels of the first
-                    BOLD frame.
-                --brainsignal
-                    Mask will consist of all the voxels that are of value 300 or
-                    higher in the first BOLD frame (this gave a good coarse
-                    brain mask for images intensity normalized to mode 1000 in
-                    the NIL preprocessing stream).
-                --brainmask
-                    Mask will be the actual bet extracted brain mask based on
-                    the first BOLD frame (generated using in the
-                    creatBOLDBrainMasks command).
-                --filename
-                    All the non-zero voxels in a specified volume file will be
-                    used as a mask.
-                --false
-                    No mask will be used.
-                --same
-                    Only for dilate_mask, the mask used will be the same as
-                    smoothing mask.
-
-            Cifti smoothing:
-                For cifti format images, smoothing will be run using wb_command.
-                The following parameters can be set:
-
-                --surface_smooth (float, default 2.0):
-                    FWHM for Gaussian surface smoothing in mm.
-                --volume_smooth (float, default 2.0):
-                    FWHM for Gaussian volume smoothing in mm.
-                --framework_path (str, default ''):
-                    The path to framework libraries on the Mac system. No need
-                    to use it currently if installed correctly.
-                --wb_command_path (str, default ''):
-                    The path to the wb_command executive. No need to use it
-                    currently if installed correctly.
-
-                Results:
-                    The resulting smoothed files are saved with '_s' added to
-                    the BOLD root filename.
-
-
-        Temporal filtering:
-            Temporal filtering is accomplished using img_filter nimage method.
-            The code is adopted from the FSL C++ code enabling appropriate
-            handling of bad frames (as described above - see SCRUBBING). The
-            parameters are:
-
-            --hipass_filter (float, default 0.008):
-                The frequency for high-pass filtering in Hz.
-            --lopass_filter (float, default 0.09):
-                The frequency for low-pass filtering in Hz.
-
-            Please note that the values finally passed to img_filter method are
-            the respective sigma values computed from the specified frequencies
-            and TR.
-
-            Filtering of nuisance signal, movement, task, and events
-            Besides data, nuisance signal, motion parameters, and event
-            regressors can be filtered as well. What to filter beside data can
-            be specified by a comma separated list using the following
-            parameters:
-
-            --hipass_do (str, default 'nuisance')
-                What to high-pass filter besides data – options are: nuisance,
-                movement, events, task. Default is 'nuisance'.
-
-            --lopass_do (str, default 'nuisance,movement,task,events')
-                What to lo-pass filter besides data – options are: nuisance,
-                movement, events, task. Default is 'nuisance, movement, task,
-                events'.
-
-            Note that 'events' refers to regressors created based on events
-            as specified in the fidl file, whereas 'task' refers to a task
-            matrix that is passed directy in the matlab function call.
-
-            Results:
-                The resulting filtered files are saved with '_hpss' or '_bpss'
-                added to the BOLD root filename for high-pass and low-pass
-                filtering, respectively.
-
-        Regression:
-            Regression is a complex step in which GLM is used to estimate the
-            beta weights for the specified nuisance regressors and events. The
-            resulting beta weights are then stored in a GLM file (a regular
-            file with additional information on the design used) and residuals
-            are stored in a separate file. This step can therefore be used for
-            two purposes: (1) to remove nuisance signal and event structure
-            from BOLD files, removing unwanted potential sources of correlation
-            for further functional connectivity analyses, and (2) to get task
-            beta estimates for further activation analyses. The following
-            parameters are used in this step:
-
-            --bold_nuisance (str, default 'm,m1d,mSq,m1dSq,V,WM,WB,1d'):
-                A comma separated list of regressors to include in GLM.
-                Possible values are:
-
-                - 'm'     ... motion parameters
-                - 'm1d'   ... first derivative of motion parameters
-                - 'mSq'   ... squared motion parameters
-                - 'm1dSq' ... squared first derivative of motion parameters
-                - 'V'     ... ventricles signal
-                - 'WM'    ... white matter signal
-                - 'WB'    ... whole brain signal
-                - '1d'    ... first derivative of above nuisance signals
-                - 'e'     ... events listed in the provided fidl files.
-
-            --event_string (str, default ''):
-                A string describing, how to model the events listed in the
-                provided fidl files.
-            --glm_matrix (str, default 'none'):
-                Whether to save the GLM matrix as a text file ('text'), a png
-                image file ('image'), both ('both') or not ('none').
-            --glm_results (str, default 'c,r')
-                A string  which of the GLM analysis results are saved.
-                Possible values are:
-
-                - 'c'   ... Saving of resulting beta coefficients.
-                - 'z'   ... Saving of resulting z-scores of beta coefficients.
-                - 'p'   ... Saving of session-level coefficient p-values.
-                - 'se'  ... Saving of standard errors of beta coefficients.
-                - 'r'   ... Saving of resulting residuals of the GLM.
-                - 'all' ... Saving all of the results above.
-
-            --glm_name (str, default ''):
-                An additional name to add to the residuals and GLM files to
-                distinguish between different possible models used.
-
-            GLM modeling:
-                There are two important variables that affect the exact GLM
-                model used to estimate nuisance and task beta coefficients and
-                regress them from the signal. The first is the optional number
-                following the 'r' command in the --bold_actions parameter.
-                There are three options:
-
-                - '0' ... Estimate nuisance regressors for each bold file
-                  separately, however, model events across all bold files (the
-                  default if no number is) specified.
-                - '1' ... Estimate both nuisance regressors and task regressors
-                  for each bold run separately.
-                - '2' ... Estimate both nuisance regressors as well as task
-                  regressors across all bold runs.
-
-                The second key variable is the event string provided by the
-                --event_string parameter. The event string is a pipe ('|')
-                separated list of regressor specifications. The possibilities
-                are discussed below:
-
-                Unassumed Modeling:
-                    ::
-
-                        <fidl code>:<length in frames>
-
-                    Where <fidl code> is the code for the event used in the fidl
-                    file, and <length in frames> specifies, for how many frames
-                    of the bold run (since the onset of the event) the event
-                    should be modeled.
-
-                Assumed Modeling:
-                    ::
-
-                        <fidl code>:<hrf>[-run|-uni][:<length>]
-
-                    Where <fidl code> is the same as above, <hrf> is the type of
-                    the hemodynamic response function to use, '-run' and '-uni'
-                    specify how the regressor should be normalized, and
-                    <length> is an optional parameter, with its value dependent
-                    on the model used. The allowed <hrf> are:
-
-                    - 'boynton' ... uses the Boynton HRF
-                    - 'SPM'     ... uses the SPM double gaussian HRF
-                    - 'u'       ... unassumed (see above)
-                    - 'block'   ... block response.
-
-                    For the first two, the <length> parameter is optional and
-                    would override the event duration information provided in
-                    the fidl file. For 'u' the length is the same as in
-                    previous section: the number of frames to model. For
-                    'block' length should be two numbers separated by a colon
-                    (e.g. 2:9) that specify the start and end offset (from the
-                    event onset) to model as a block.
-
-                    Assumed HRF regressors normalization:
-                        hrf_types `boynton` and `SPM` can be marked with an
-                        additional flag denoting how to normalize the
-                        regressor.
-
-                        In case of `<hrf function>-uni`, e.g. 'boynton-uni' or
-                        'SPM-uni', the HRF function will be normalized to have
-                        the area under the curve equal to 1. This ensures
-                        uniform and universal, scaling of the resulting
-                        regressor across all event lengths. In addition, the
-                        scaling is not impacted by weights (e.g. behavioral
-                        coregressors), which in turn ensures that the weights
-                        are not scaled.
-
-                        In case of `<hrf function>-run`, e.g. `boynton-run` or
-                        `SPM-run`, the resulting regressor is normalized to
-                        amplitude of 1 within each bold run separately. This
-                        can result in different scaling of regressors with
-                        different durations, and of the same regressor across
-                        different runs. Scaling in this case is performed after
-                        the signal is weighted, so in effect the scaling of
-                        weights (e.g. behavioral regressors), can differ across
-                        bold runs.
-
-                        The flag can be abbreviated to '-r' and '-u'. If not
-                        specified, '-run' will be assumed (the default might
-                        change).
-
-                Naming And Behavioral Regressors:
-                    Each of the above (unassumed and assumed modeling
-                    specification) can be followed by a ">" (greater-than
-                    character), which signifies additional information in the
-                    form::
-
-                        <name>[:<column>[:<normalization_span>[:<normalization_method>]]]
-
-                    --name
-                        The name of the resulting regressor.
-                    --column
-                        The number of the additional behavioral regressors
-                        column in the fidl file (1-based) to use as a weight
-                        for the regressors.
-                    --normalization_span
-                        Whether to normalize the behavioral weight within a
-                        specific event type ('within') or across all events
-                        ('across'). [within]
-                    --normalization_method
-                        The method to use for normalization. Options are:
-
-                        - 'z'    (compute Z-score)
-                        - '01'   (normalize to fixed range 0 to 1)
-                        - '-11'  (normalize to fixed range -1 to 1)
-                        - 'none' (use weights as provided in fidl file)
-
-                    Example string::
-
-                        'block:boynton|target:9|target:9>target_rt:1:within:z'
-
-                    This would result in three sets of task regressors: one
-                    assumed task regressor for the sustained activity across
-                    the block, one unassumed task regressor set spanning 9
-                    frames that would model the presentation of the target, and
-                    one behaviorally weighted unassumed regressor that would
-                    for each frame estimate the variability in response as
-                    explained by the reaction time to the target.
-
-    Examples:
-        Activation analysis::
-
-            qunex preprocess_conc \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --parsessions=10 \\
-                --bolds=SRT \\
-                --event_file=SRT \\
-                --glm_name=-M1 \\
-                --bold_actions="s,r,c" \\
-                --bold_nuisance=e \\
-                --mov_bad=none \\
-                --event_string="block:boynton|target:9|target:9>target_rt:1:within:z" \\
-                --glm_matrix=both \\
-                --glm_residuals=none \\
-                --pignore="hipass=keep|regress=keep|lopass=keep"
-
-        Functional connectivity preprocessing::
-
-            qunex preprocess_conc \\
-                --batchfile=fcMRI/sessions_hcp.txt \\
-                --sessionsfolder=sessions \\
-                --overwrite=no \\
-                --parsessions=10 \\
-                --bolds=SRT \\
-                --event_file=SRT \\
-                --glm_name=-FC \\
-                --bold_actions="s,h,r,c,l" \\
-                --bold_nuisance="m,V,WM,WB,1d,e" \\
-                --mov_bad=udvarsme \\
-                --event_string="block:boynton|target:9" \\
-                --glm_matrix=none \\
-                --glm_residuals=save \\
-                --pignore="hipass=linear|regress=ignore|lopass=linear"
-    """
-    log = ReportLog()
-
-    pc.do_options_check(options, sinfo, "preprocess_conc")
-
-    log.rule()
-    log.info(f"Session id: {sinfo['id']} \n[started on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}]")
-    log.action("Running", "Preprocessing conc bundles ...", options["run"], level="info")
-    log.info(f"Files in 'images{options['img_suffix']}/functional{options['bold_variant']} will be processed.")
-
-    # --- extract conc and fidl names
-    concs = [e.strip().replace(".conc", "") for e in options["bolds"].split("|")]
-    fidls = [e.strip().replace(".fidl", "") for e in options["event_file"].split("|")]
-
-    # --- define the tail
-    options["bold_tail"] = options["nifti_tail"]
-    if options["image_target"] in ["cifti", "dtseries", "ptseries"]:
-        options["bold_tail"] = options["cifti_tail"]
-
-    concroot = "_".join(
-        e
-        for e in [
-            options["boldname"] + options["bold_tail"],
-            options["image_target"].replace("cifti", "dtseries"),
-        ]
-        if e
-    )
-    report = ""
-
-    failed = 0
-    if len(concs) != len(fidls):
-        log.error(f"Number of conc files ({len(concs)}) does not match number of event files ({len(fidls)}), processing aborted!")
-
-    else:
-        for nb in range(0, len(concs)):
-            tconc = concs[nb]
-            tfidl = fidls[nb]
-            options["concname"] = tconc
-            if tfidl:
-                options["fidlname"] = tfidl
-            else:
-                options["fidlname"] = ""
-
-            try:
-                log.raw(f"\n\nConc bundle: {tconc}")
-
-                d = pc.get_session_folders(sinfo, options)
-                f = {}
-                f_conc = os.path.join(
-                    d["s_bold_concs"], concroot + "_" + tconc + ".conc"
-                )
-                f_fidl = os.path.join(d["s_bold_events"], tfidl + ".fidl")
-
-                # --- find conc data
-
-                if overwrite or not os.path.exists(f_conc):
-                    tf = pc.find_file(sinfo, options, tconc + ".conc")
-                    if tf:
-                        log.detail(f"getting conc data from {tf}")
-                        if os.path.exists(f_conc):
-                            _remove(log, options, f_conc)
-                        _copy(log, options, tf, f_conc)
-
-                    else:
-                        log.error(f"Conc data file ({tconc}) does not exist in the expected locations! Skipping this conc bundle.", depth=1)
-                        failed += 1
-                        continue
-                else:
-                    log.detail("conc data present")
-
-                # --- find fidl data
-
-                if "e" in [e.strip() for e in options["bold_nuisance"].split(",")]:
-                    if overwrite or not os.path.exists(f_fidl):
-                        tf = pc.find_file(sinfo, options, tfidl + ".fidl")
-                        if tf:
-                            log.detail(f"getting event data from {tf}")
-                            if os.path.exists(f_fidl):
-                                _remove(log, options, f_fidl)
-                            _copy(log, options, tf, f_fidl)
-                        else:
-                            log.error(f"Event data file ({tfidl}) does not exist in the expected locations! Skipping this conc bundle.", depth=1)
-                            failed += 1
-                            continue
-                    else:
-                        log.detail("event data present")
-                else:
-                    log.detail(f"event data not needed (e not specified in --bold_nuisance) {options['bold_nuisance']}")
-
-                # --- loop through bold files
-
-                conc = gi.read_conc(f_conc, boldname=options["boldname"])
-                nconc = []
-                bolds = []
-                rstatus = True
-                check = {"ok": [], "bad": []}
-
-                if len(conc) == 0:
-                    log.error(f"No valid image files in conc file ({f_conc})! Skipping this conc bundle.", depth=1)
-                    failed += 1
-                    continue
-
-                for c in conc:
-                    # print("c from conc:", c)
-                    boldnum = c[1]
-                    boldname = options["boldname"] + boldnum
-                    bolds.append(boldnum)
-
-                    # if absolute path flag use session folder from conc file
-                    if options["conc_use"] == "absolute":
-                        # extract session folder from conc file
-                        options["sessionsfolder"] = (c[0].split(sinfo["id"]))[0]
-                        d["s_base"] = options["sessionsfolder"] + sinfo["id"]
-                        options["bold_tail"] = (c[0].split(boldname))[1].replace(
-                            pc.get_extension(options["image_target"]), ""
-                        )
-
-                    log.raw("\n\nLooking up: " + boldname + " ...")
-
-                    # --- filenames
-                    f = pc.get_file_names(sinfo, options)
-                    f.update(pc.get_bold_file_names(sinfo, boldname, options))
-
-                    # if absolute path flag use also exact filename (extension)
-                    if options["conc_use"] == "absolute":
-                        f["bold"] = c[0]
-
-                    # --- check for data availability
-
-                    # r += '\n    ... checking for data'
-                    status = True
-
-                    # --- bold
-                    status = pc.check_for_file(f["bold"],
-                        "bold data present",
-                        f"bold data missing [{f['bold']}]",
-                        status=status,
-                        _log=log,
-                    )
-                    nconc.append((f["bold"], boldnum))
-
-                    # --- movement
-                    if "r" in options["bold_actions"] and (
-                        "m" in options["bold_nuisance"]
-                        or "m" in options["bold_actions"]
-                    ):
-                        status = pc.check_for_file(f["bold_mov"],
-                            "movement data present",
-                            f"movement data missing [{f['bold_mov']}]",
-                            status=status,
-                            _log=log,
-                        )
-
-                    # --- bold stats
-                    if "m" in options["bold_actions"]:
-                        status = pc.check_for_file(f["bold_stats"],
-                            "bold statistics data present",
-                            f"bold statistics data missing [{f['bold_stats']}]",
-                            status=status,
-                            _log=log,
-                        )
-
-                    # --- bold scrub
-                    if any(
-                        [
-                            e in options["pignore"]
-                            for e in ["linear", "spline", "ignore"]
-                        ]
-                    ):
-                        status = pc.check_for_file(f["bold_scrub"],
-                            "bold scrubbing data present",
-                            f"bold scrubbing data missing [{f['bold_scrub']}]",
-                            status=status,
-                            _log=log,
-                        )
-
-                    # --- check for nuisance data files if doing regression
-
-                    if "r" in options["bold_actions"] and any(
-                        [e in options["bold_nuisance"] for e in ["V", "WM", "WB"]]
-                    ):
-                        status = pc.check_for_file(f["bold_nuisance"],
-                            "bold nuisance signal data present",
-                            f"bold nuisance signal data missing [{f['bold_nuisance']}]",
-                            status=status,
-                            _log=log,
-                        )
-
-                    # --- check
-                    if not status:
-                        log.error("Files missing!")
-                        rstatus = False
-                        check["bad"].append(boldnum)
-                    else:
-                        check["ok"].append(boldnum)
-
-                # --- all files reviewed continuing conc processing
-
-                report = ""
-                if len(check["ok"]) > 0:
-                    report += "%d bolds ok [%s], " % (
-                        len(check["ok"]),
-                        " ".join(check["ok"]),
-                    )
-                else:
-                    report += "0 bolds ok, "
-
-                if len(check["bad"]) > 0:
-                    report += "%d bold not ok [%s], " % (
-                        len(check["bad"]),
-                        " ".join(check["bad"]),
-                    )
-                else:
-                    report += "0 bolds not ok"
-
-                if not rstatus:
-                    log.error("Due to missing data we are skipping this conc bundle!")
-                    report += " => missing data"
-                    failed += 1
-                    continue
-
-                if options["run"] == "run":
-                    gi.write_conc(f_conc, nconc)
-                else:
-                    log.detail(f"test, not rewritten: {f_conc}")
-
-                # --- run matlab preprocessing script
-
-                if overwrite:
-                    boldow = "true"
-                else:
-                    boldow = "false"
-
-                done = f["conc_final"] + ".ok"
-
-                scrub = (
-                    "radius:%(mov_radius)d|fdt:%(mov_fd).2f|dvarsmt:%(mov_dvars).2f|dvarsmet:%(mov_dvarsme).2f|after:%(mov_after)d|before:%(mov_before)d|reject:%(mov_bad)s"
-                    % (options)
-                )
-                opts = (
-                    "boldname=%(boldname)s|fidlname=%(fidlname)s|concname=%(concname)s|surface_smooth=%(surface_smooth)f|volume_smooth=%(volume_smooth)f|voxel_smooth=%(voxel_smooth)f|hipass_filter=%(hipass_filter)f|lopass_filter=%(lopass_filter)f|framework_path=%(framework_path)s|wb_command_path=%(wb_command_path)s|smooth_mask=%(smooth_mask)s|dilate_mask=%(dilate_mask)s|glm_matrix=%(glm_matrix)s|glm_residuals=%(glm_residuals)s|glm_results=%(glm_results)s|glm_name=%(glm_name)s|bold_tail=%(bold_tail)s|ref_bold_tail=%(nifti_tail)s|bold_variant=%(bold_variant)s|img_suffix=%(img_suffix)s"
-                    % (options)
-                )
-
-                mcomm = (
-                    "fc_preprocess_conc('%s', [%s], '%s', %.3f,  %d, '%s', [], '%s.fidl', '%s', '%s', %s, '%s', '%s', '%s', '%s', '%s')"
-                    % (
-                        d["s_base"],  # --- session folder
-                        " ".join(
-                            bolds
-                        ),  # --- vector of bold runs in the order of the conc file
-                        options[
-                            "bold_actions"
-                        ],  # --- which steps to perform in what order (s, h, r0/r1/r2, c, p, l)
-                        options["tr"],  # --- TR
-                        options[
-                            "omit"
-                        ],  # --- the number of frames to omit at the start of each run
-                        options[
-                            "bold_nuisance"
-                        ],  # --- nuisance regressors (m, v, wm, wb, d, t, e)
-                        tfidl,  # --- event file to be used for task regression (w/o .fidl)
-                        options[
-                            "event_string"
-                        ],  # --- event string specifying what and how of task to regress
-                        options[
-                            "bold_prefix"
-                        ],  # --- optional prefix to the resulting bolds
-                        boldow,  # --- whether to overwrite existing files
-                        gi.get_img_format(
-                            f["bold_final"]
-                        ),  # --- the format of the images (.nii vs. .4dfp.img)
-                        scrub,  # --- scrub parameters string
-                        options[
-                            "pignore"
-                        ],  # --- how to deal with bad frames ('hipass:keep/linear/spline|regress:keep/ignore|lopass:keep/linear/spline')
-                        opts,  # --- additional options
-                        done,
-                    )
-                )  # --- file to save when done
-
-                comm = (
-                    '%s "try %s; catch ME, general_report_crash(ME); exit(1), end; exit;"'
-                    % (mcommand, mcomm)
-                )
-
-                log.blank()
-                log.action(
-                    "Running", "nuisance and task removal", options["run"], level="info"
-                )
-                if options["print_command"] == "yes":
-                    log.raw("\n" + comm + "\n")
-                if options["run"] == "run":
-                    # the third return is this call's failure flag, not the
-                    # count of failed conc bundles this loop is accumulating
-                    endlog, status, _ = pc.run_external_for_file(
-                        done,
-                        comm,
-                        "running matlab/octave fc_preprocess_conc on bolds [%s]"
-                        % (" ".join(bolds)),
-                        overwrite=overwrite,
-                        thread=sinfo["id"],
-                        remove=options["log"] == "remove",
-                        task=options["command_ran"],
-                        logfolder=options["comlogs"],
-                        logtags=[
-                            options["bold_variant"],
-                            options["bolds"],
-                            options["glm_name"],
-                            options["logtag"],
-                        ],
-                        shell=True,
-                        _log=log,
-                    )
-                    status = pc.check_for_file(
-                        done,
-                        bad=f"Matlab/Octave has failed preprocessing BOLD using command: \n---> {mcomm}\n",
-                        bad_level="error",
-                        _log=log,
-                    )
-                    if os.path.exists(done):
-                        os.remove(done)
-                    if status:
-                        report += " => processed ok"
-                    else:
-                        report += " => processing failed"
-                        failed += 1
-                else:
-                    log.detail(f"test, not run: {comm}", depth=1)
-                    if os.path.exists(done):
-                        report += " => already done"
-                    else:
-                        report += " => ready"
-                        failed += 1
-
-            except ge.CommandFailed as e:
-                log.raw("\n" + ge.report_command_failed("preprocess_conc", e))
-                report += " => processing failed"
-                failed += 1
-            except ge.CommandError as e:
-                log.raw("\n" + ge.report_command_error("preprocess_conc", e))
-                report += " => processing failed"
-                failed += 1
-            except (pc.ExternalFailed, pc.NoSourceFolder) as errormessage:
-                log.raw(str(errormessage))
-                report += " => processing failed"
-                failed += 1
-            except Exception:
-                report += " => processing failed"
-                log.error(f"Unknown error occured: \n...................................\n{traceback.format_exc()}...................................\n")
-                time.sleep(5)
-                failed += 1
-
-    log.blank()
-    log.info(f"Conc preprocessing (v2) completed on {datetime.now().strftime('%A, %d. %B %Y %H:%M:%S')}")
-    log.rule()
-
-    # print r
-    return log.result(report, failed, sinfo["id"])
